\name{AcademyAwards}
\alias{AcademyAwards}
\docType{data}
\title{
AcademyAwards
}
\description{
Data on Academy Awards for a sample of movies
}
%\useage{data(AcademyAwards)}
\format{
  A dataset with 69 observations on the following 12 variables.
  \tabular{rl}{
    \code{FilmTitle} \tab {Name of movie}\cr
    \code{PreviousAwards} \tab {Number of previous Academy Awards for the cast}\cr
    \code{Nominations} \tab {Number of Academy Award nominations}\cr
    \code{ActorFilms} \tab {Number of previous films by lead actor}\cr
    \code{ActressFilms} \tab {Number of previous films by lead actress}\cr
    \code{Length} \tab {Running time for movie (in minutes)}\cr
    \code{Budget} \tab {Production cost (millions)}\cr
    \code{PctPositiveReviews} \tab {Percent of positive reviews}\cr
    \code{OscarsWon} \tab {Number of Academy Awards won}\cr
    \code{BestPicture} \tab {Best picture award?  1=yes  or  0=no}\cr
    \code{OpenGross} \tab {Opening weekend gross (millions)}\cr
    \code{Gross} \tab {Worldwide gross (millions)}\cr
  }
}
\details{
Data for a sample of movies.
}
\source{
??
}

\keyword{datasets}
