\name{AlcoholResponse}
\alias{AlcoholResponse}
\docType{data}
\title{
AlcoholResponse
}
\description{
Reaction times in a driving simulation after drinking alcohol or a placebo.
}
\usage{data(AlcoholResponse)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{Response}}{Reaction time (seconds)}
    \item{\code{Treatment}}{Type of drink: \code{Placebo} or \code{Alcohol}}
  }
}
\details{
Participants were randomly assigned to consume either an alcoholic or nonalcoholic (placebo) drink
before their mean reaction time was measured for a series of simulated driving situations. 
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
%% data(AlcoholResponse)
}
\keyword{datasets}
