\name{AlcoholResponse2}
\alias{AlcoholResponse2}
\docType{data}
\title{
AlcoholResponse2
}
\description{
Reaction times in a driving simulation after drinking alcohol or a placebo.
}
%\useage{data(AlcoholResponse2)}
\format{
  A dataset with 10 observations on each of the following 2 variables.
  \tabular{rl}{
    \code{Placebo} \tab {Reaction time (in seconds) for participants in the placebo group}\cr
    \code{Alcohol} \tab {Reaction time (in seconds) for participants in the alcohol group}\cr
  }
}
\details{
Participants were randomly assigned to consume either an alcoholic or nonalcoholic (placebo) drink
before their mean reaction time was measured for a series of simulated driving situations. 
}
\source{
???
}

\keyword{datasets}
