\name{Amyloid}
\alias{Amyloid}
\docType{data}
\title{Amyloid-beta and Cognitive Impairment}
\description{
Amyloid-beta and cognitive impairment for a sample of Catholic priests
}
\format{
  A data frame with 57 observations on the following 2 variables.
  \describe{
    \item{\code{Group}}{ \code{mAD}=Alzheimer's, \code{MCI}=mild impairment, \code{NCI}=no impairment}
    \item{\code{Abeta}}{Amount of Abeta from the posterior cingulate cortex (pmol/g tissue)}
  }
}
\details{
Amyloid-beta (Abeta) is a protein fragment that has been linked to Alzheimer's disease. Autopsies from a sample of Catholic priests included measurements of Abeta (pmol/g tissue from the posterior cingulate cortex) from three groups: subjects who had exhibited no cognitive impairment before death, subjects who had exhibited mild cognitive impairment, and subjects who had mild to moderate Alzheimer's disease.
}
\source{
Violetta N. Pivtoraiko, Eric E. Abrahamson, Sue E. Leurgans, Steven T. DeKosky,
Elliott J. Mufson,, Milos D. Ikonomovic (2015)  Cortical pyroglutamate amyloid-beta levels and cognitive decline in Alzheimer's disease.  Neurobiology of Aging (36) 12-19.
Data are read from Figure 1, panel d.
}
\keyword{datasets}
