\name{AppleStock}
\alias{AppleStock}
\docType{data}
\title{AppleStock}
\description{
Daily price of Apple stock
}
\format{
  A data frame with 66 observations on the following 4 variables.
  \describe{
    \item{\code{Date}}{Date as mm/dd/yyy}
    \item{\code{Price}}{Closing price of Apple stock}
    \item{\code{Change}}{Change in price from previous day}
    \item{\code{Volume}}{Number of shares traded i(in millons)}
  }
}
\details{
Closing price of Apple stock (AAPL) for each trading day in a three month period from 7/21/2016 to 10/21/2016 as well as the change in stock price and number of shares traded.
}
\source{
Data downloaded from Nasdaq historical prices at  http://www.nasdaq.com/symbol/aapl/historical
}
\keyword{datasets}
