\name{ArcheryData}
\alias{ArcheryData}
\docType{data}
\title{Scores in an Archery Class}
\description{
Score results from an archery class
}
%\useage{data(ArcheryData)}
\format{
  A dataset with 18 observations on the following 7 variables.
  \tabular{rl}{
    \code{Attendance} \tab {Number of days in class}\cr
    \code{Average} \tab {Average score over all days}\cr
    \code{Sex} \tab {Coded as \code{f} or \code{m}}\cr
    \code{Day1} \tab {Archery score on first day}\cr
    \code{LastDay} \tab {Archery score on last day}\cr
    \code{Improvement} \tab {Last day - first day score}\cr
    \code{Improve} \tab {1=improved or 0=did not improve}\cr
  }
}
\details{
In 2002, Heather Tollerud, a Saint Olaf College student, undertook a study of the archery scores of
students at the college who were enrolled in an archery course. Students taking the course record a
score for each day they attend class from the first until the last day. Hopefully the instruction they
receive helps them to improve their game.
}
\source{
Student project
}

\keyword{datasets}
