\name{AudioVisual}
\alias{AudioVisual}
\docType{data}
\title{Reaction Times to Audio and Visual Stimuli}
\description{
Data from an experiment on reaction times to audio or visual stimuli by Oberlin College students.
}
\format{
  A data frame with 72 observations on the following 4 variables.
  \describe{
    \item{\code{Subject}}{SubjectIDs coded \code{s1} to \code{s36}}
    \item{\code{ResponseTime}}{Time to respond to a stimulus (in ms)}
    \item{\code{Stimulus}}{Type of stimulus  (\code{auditory} or \code{visual})}
    \item{\code{Group}}{\code{Musician} or \code{NonMusician}}
  }
}
\details{
Subjects in a reaction time study were asked to press a button as fast as possible after being exposed to either an auditory stimulus (a burst of white noise) or a visual stimulus (a circle flashing on a computer screen).  Average reaction times (ms) were recorded for between 10 and 20 trials for each type of stimulus for each subject. Data also identifies which subjects are musicians.
}
\source{
Arjuna Pettit, Jr. and Jeremy Potterfield at Oberlin College
}
\keyword{datasets}
