\name{AutoPollution}
\alias{AutoPollution}
\docType{data}
\title{
Noise Levels of Filters to Reduce Automobile Pollution
}
\description{
Measurements of noise levels for different filters to reduce pollution levels of automobiles.
}
\format{
  A dataset with 36 observations on the following 4 variables.
  \tabular{rl}{
    \code{Noise} \tab {Noise level (decibels)}\cr
    \code{Size} \tab {Vehicle size: \code{1}=small, \code{2}=medium, or \code{3}=large}\cr
    \code{Type} \tab {\code{1}=standard filter or \code{2}=new filter}\cr
    \code{Side} \tab {Side of vehicle: code{1}=right or \code{2}=left}\cr
  }
}
\details{
In a 1973 testimony before the Air and Water Pollution Subcommittee
of the Senate Public Works Committee, John McKinley, President of Texaco discussed a new filter
that had been developed to reduce pollution. Questions were raised about the effects of this filter
on other measures of vehicle performance. The data set AutoPollution gives the results of an
experiment on 36 different cars. The cars were randomly assigned to get either this new filter or a
standard filter and the noise level for each car was measured.
}
\source{
Data explanation and link can be found at http://lib.stat.cmu.edu/DASL/Stories/airpollutionfilters.html.
}
\references{
 A.Y. Lewin and M.F. Shakun, Policy Sciences: Methodology and Cases, Pergammon Press, 1976, p 313.
}

\keyword{datasets}
