\name{Backpack}
\alias{Backpack}
\docType{data}
\title{Weights of College Student Backpacks}
\description{
Backpack weights for a sample of college students
}
\format{
  A data frame with 100 observations on the following 9 variables.
  \tabular{rl}{
    \code{BackpackWeight} \tab {Backpack weight (in pounds)}\cr
    \code{BodyWeight} \tab {Body weight (in pounds)}\cr
    \code{Ratio} \tab {BackpackWeight/BodyWeight}\cr
    \code{BackProblems} \tab {\code{0}=no or \code{1}=yes}\cr
    \code{Major} \tab {Code for academic major}\cr
    \code{Year} \tab {Year in school}\cr
    \code{Sex} \tab {a factor with levels \code{Female} \code{Male}}\cr
    \code{Status} \tab {Graduate or undergraduate? \code{G} or \code{U}}\cr
    \code{Units} \tab {Number of credits taken that quarter}\cr
  }
}
\details{
A survey of students at California Polytechnic State University (San Luis Obispo) collected data
to investigate the question of whether back aches might be due to carrying heavy backpacks,
}
\source{
Mintz J., Mintz J., Moore K., and Schuh K., "Oh, My Aching Back! A Statistical Analysis of Backpack Weights,"  Stats: The Magazine for Students of Statistics, vol. 32, 2002, pp. 1719.
}

\keyword{datasets}
