\name{BaseballTimes}
\alias{BaseballTimes}
\docType{data}
\title{BaseballTimes}
\description{
Game times and boxscore information for baseball games
}
%\useage{data(BaseballTimes)}
\format{
  A dataset with 15 observations on the following 7 variables.
  \tabular{rl}{
    \code{Game} \tab {Code for opposing teams}\cr
    \code{League} \tab {\code{AL}= American League or  \code{NL}=Mational League}\cr
    \code{Runs} \tab {Total number of runs scored (both teams)}\cr
    \code{Margin} \tab {Margin of victory (Winner-Loser score)}\cr
    \code{Pitchers} \tab {Total number of pitchers used (both teams}\cr
    \code{Attendance} \tab {Number of spectators at the game}\cr
    \code{Time} \tab {Total time for the game (in minutes)}\cr
  }
}
\details{
Data were collected for 15 Major Leage Baseball (MLB) games played on August 26, 2008. 
}
\source{
Data from boxscores at www.baseball-reference.com
}

\keyword{datasets}
