\name{BaseballTimes2017}
\alias{BaseballTimes2017}
\docType{data}
\title{Baseball Game Times of One Day in 2017}
\description{
Times for one day's major league baseball games
}
\format{
  A data frame with 14 observations on the following 7 variables.
  \describe{
    \item{\code{Game}}{MLB teams that played}
    \item{\code{League}}{\code{AL}=American League, \code{IL}=Interleague, or \code{NL}=National League}
    \item{\code{Runs}}{Runs scored by the two teams combined}
    \item{\code{Margin}}{Winning margin}
    \item{\code{Pitchers}}{Number of pitchers used total for two teams}
    \item{\code{Attendance}}{Announced attendance}
    \item{\code{Time}}{Time in minutes to play the game}
  }
}
\details{
 Data from all MLB games played on August 11, 2017. There were no extra-innings game nor any rain delays.
 }
\source{
https://www.baseball-reference.com/boxes/?month=8&day=11&year=2017
}
\keyword{datasets}
