\name{BeeStings}
\alias{BeeStings}
\docType{data}
\title{Do Bee Stings Depend on Previous Stings?}
\description{
Data from an experiment to see it the number of bee stings depends on previous stings.
}
\format{
  A data frame with 18 observations on the following 3 variables.
  \tabular{rl}{
    \code{Occasion} \tab {Trial: I to IX}\cr
    \code{Treatment} \tab {\code{Fresh} or \code{Stung}}\cr
    \code{Stingers} \tab {Number of stingers}\cr
  }
}
\details{
If you are stung by a bee, does that make you more likely to get stung again? Might bees leave
behind a chemical message that tells other bees to attack you? To test this hypothesis, scientists
dangled a 4x4 array of 16 muslin-wrapped cotton balls over a beehive.
Eight of 16 balls had been previously stung; the other eight were
fresh. The response was the total number of new stingers left behind by the bees.
The process was repeated for a total of nine trials.

Used in first edition, but not second edition.
}
\source{
Free, J.B. (1961) "The stinging response of honeybees," Animal Behavior, Vol. 9, pp 193-196.
}

\keyword{datasets}
