\name{BirdCalcium}
\alias{BirdCalcium}
\docType{data}
\title{Effect of a Hormone on Bird Calcium Levels}
\description{
An experiment on the effects of a hormone on blood calcium levels in robins
}
\format{
  A data frame with 20 observations on the following 5 variables.
  \describe{
    \item{\code{Bird}}{ID number for each bird (1 to 20)}
    \item{\code{Sex}}{\code{female} or \code{male}}
    \item{\code{Hormone}}{Treated with hormone (\code{no} or \code{yes})}
    \item{\code{Group}}{Combined Sex and Hormone (\code{F No}, \code{F Yes}, \code{M No}, or \code{M Yes})}
    \item{\code{Ca}}{Blood calcium level (mg per 100 ml)}
  }
}
\details{
An experiment looked at the effects of treatment with a hormone for increasing the concentration of calcium in birds.  Twenty birds (robins) were used in the study, ten male and ten female, equally divided between the hormone and no hormone treatments.
}
\source{
Bliss, Chester (1970), Statistics in Biology, McGraw-Hill
}
\keyword{datasets}
