\name{Blood1}
\alias{Blood1}
\docType{data}
\title{Blood Pressure, Weight, and Smoking Status}
\description{
Systolic blood pressure, weight and smoking status for a sample of 500 adults
}
\format{
  A data frame with 500 observations on the following 3 variables.
  \tabular{rl}{
    \code{SystolicBP} \tab {Systolic blood pressure (mm of Hg)}\cr
    \code{Smoke} \tab {\code{Y}=smoker or \code{N}=non-smoker}\cr
    \code{Overwt} \tab {\code{1}=normal, \code{2}=overweight, or \code{3}=obese}\cr
  }
}
\details{
Data on systolic blood pressure, along with smoker status and weight status, for a sample of 500 adults.
}
\source{
Data are part of a larger case study for the 2003 Annual Meeting of the Statistical Society of Canada. \cr
http://www.ssc.ca/en/education/archived-case-studies/case-studies-for-the-2003-annual-meeting-blood-pressure.
}

\keyword{datasets}
