\name{BlueJays}
\alias{BlueJays}
\docType{data}
\title{Blue Jay Measurements}
\description{
Body measurements for a sample of blue jays
}
\format{
  A data frame with 123 observations on the following 9 variables.
  \tabular{rl}{
    \code{BirdID} \tab {ID tag for bird}\cr
    \code{KnownSex} \tab {Sex coded as \code{F} or \code{M}}\cr
    \code{BillDepth} \tab {Thickness of the bill measured at the nostril (in mm)}\cr
    \code{BillWidth} \tab {Width of the bill (in mm)}\cr
    \code{BillLength} \tab {Length of the bill (in mm)}\cr
    \code{Head} \tab {Distance from tip of bill to back of head (in mm)}\cr
    \code{Mass} \tab {Body mass (in grams)}\cr
    \code{Skull} \tab {Distance from base of bill to back of skull (in mm)}\cr
    \code{Sex} \tab {Sex coded as 0=female or 1=male}\cr
  }
}
\details{
Body measurements for captured blue jays.  Values are averaged for birds captured more than once.
}
\source{
Data from Keith Tarvin, Department of Biology, Oberlin College
}

\keyword{datasets}
