\name{BreesPass}
\alias{BreesPass}
\docType{data}
\title{Drew Brees Passing Statistics (2016)}
\description{
Passing statistics for football quarterback Drew Brees in 2016
}
\format{
  A data frame with 16 observations on the following 5 variables.
  \describe{
    \item{\code{Game}}{Game number (1 is the first game of the regular season)}
    \item{\code{Opponent}}{Opponent abbreviation}
    \item{\code{Completed}}{Number of completed passes}
    \item{\code{Attempts}}{Pass attempts}
    \item{\code{Yards}}{Passing yards}
  }
}
\details{
Drew Brees was the quarterback for the NFL's New Orleans Saints football team in 2016.  This dataset shows some of his passing statistics for each of the 16 regular season games.
}
\source{
http://www.espn.com/nfl/player/gamelog/_/id/2580/year/2016
}
\keyword{datasets}
