\name{BritishUnions}
\alias{BritishUnions}
\docType{data}
\title{Attitudes Towards British Trade Unions}
\description{
Poll attitudes towards British trade unions
}
\format{
  A data frame with 17 observations on the following 7 variables.
  \tabular{rl}{
    \code{Date} \tab {Month of the poll \code{Aug-77} to \code{Sep-79}}\cr
    \code{AgreePct} \tab {Percent who agree (unions have too much power)}\cr
    \code{DisagreePct} \tab {Percent who disagree}\cr
    \code{NetSupport} \tab {DisagreePct-AgreePct}\cr
    \code{Months} \tab {Months since August 1975}\cr
    \code{Late} \tab {\code{1}=after 1986 or \code{0}=before 1986}\cr
    \code{Unemployment} \tab {Unemployment rate}\cr
  }
}
\details{
The British polling company Ipsos MORI conducted several opinion
polls in the UK between 1975 and 1995 in which they asked whether people agree or disagree
with the statement "Trade unions have too much power in Britain today".
}
\source{
Data from the Ipsos MORI website at  \cr
http://www.ipsos-mori.com/researchpublications/researcharchive/poll.aspx?oItemID=94
}

\keyword{datasets}
