\name{ButterfliesBc}
\alias{ButterfliesBc}
\docType{data}
\title{Butterfly (Boloria chariclea) Measurements}
\description{
Measurements for a sample of butterflies in Greenland
}
\format{
  A data frame with 32 observations on the following 4 variables.
  \describe{
    \item{\code{Temp}}{Average temperature for preceding summer (Celsius)}
    \item{\code{Wing}}{Average wing length (mm)}
    \item{\code{Sex}}{ \code{Female} or \code{Male}}
    \item{\code{Species}}{all are \code{Bc}, Boloria chariclea}
  }
}
\details{
Scientists measured wing length of a species of butterfly, Boloria chariclea (Bc), in Greenland each year from 1996 through 2013. They also recorded summer temperatures.
}
\source{
Digitized data from plots in Bowden, J. et al., "High-Arctic butterflies become smaller with rising temperatures", published in Biology Letters 11: 20150574
}
\keyword{datasets}
