\name{CalciumBP}
\alias{CalciumBP}
\docType{data}
\title{Do Calcium Supplements Lower Blood Pressure?}
\description{
An experiment on calcium supplements and blood pressure in 21 men
}
\format{
  A data frame with 21 observations on the following 2 variables.
  \tabular{rl}{
    \code{Treatment} \tab {\code{Calcium} or \code{Placebo}}\cr
    \code{Decrease} \tab {Beginning-ending blood pressure}\cr
  }
}
\details{
The purpose of this study was to see whether daily calcium supplements can lower blood pressure.
The subjects were 21 men; each was randomly assigned either to a treatment group or to a control
group. Those in the treatment group took a daily pill containing calcium. Those in the control
group took a daily pill with no active ingredients. Each subject's blood pressure was measured at
the beginning of the 12-week study, and again at the end.  The decrease in blood pressure (begin-end)
was recorded (so a negative value means blood pressure increased).
}
\source{
Dataset downloaded from online data source Data and Story Library, \cr
http://lib.stat.cmu.edu/DASL/Stories/CalciumandBloodPressure.html
}

\keyword{datasets}
