\name{CancerSurvival}
\alias{CancerSurvival}
\docType{data}
\title{Survival Times for Different Cancers}
\description{
Cancer survival with ascorbate supplement
}
\format{
  A dataset with 64 observations on the following 2 variables.
  \tabular{rl}{
    \code{Survival} \tab {Survival time (in days)}\cr
    \code{Organ} \tab {\code{Breast}, \code{Bronchus}, \code{Colon}, \code{Ovary}, or \code{Stomach}}\cr
  }
}
\details{
In the 1970's doctors wondered if giving terminal cancer patients a supplement of ascorbate would
prolong their lives. They designed an experiment to compare cancer patients who received ascorbate
to cancer patients who did not receive the supplement. The result of that experiment was that, in
fact, ascorbate did seem to prolong the lives of these patients. But then a second question arose.
Was the effect of the ascorbate different when different organs were affected by the cancer? The
researchers took a second look at the data. This time they concentrated only on those patients who
received the ascorbate and divided the data up by which organ was affected by the cancer. They
had 5 different organs represented among the patients (all of whom only had one organ affected):
Stomach, bronchus, colon, ovary, and breast.
}
\source{
From the article "Supplemental Ascorbate in the Supportive Treatment of Cancer: Reevaluation of Prolongation
of Survival Times in Terminal Human Cancer" by Ewan Cameron and Linus Pauling, Proceedings of the National
Academy of Sciences of the United States of America, Vol. 75, No. 9 (Sep., 1978), pp. 4538-4542.
}

\keyword{datasets}
