\name{Caterpillars}
\alias{Caterpillars}
\docType{data}
\title{Measurements of Manduca Sexta Caterpillars}
\description{
Measurements on a sample of Manduca Sexta caterpillars
}
\format{
  A data frame with 267 observations on the following 18 variables.
  \tabular{rl}{
    \code{Instar} \tab {Coded from 1 (smallest) to 5 (largest) indicating stage of the caterpillar's life}\cr
    \code{ActiveFeeding} \tab {Indicator (\code{Y} or \code{N}) of whether or not the animal is actively feeding}\cr
    \code{Fgp} \tab {Indicator (\code{Y} or \code{N}) of whether or not the animal is in a free growth period}\cr
    \code{Mgp} \tab {Indicator (\code{Y} or \code{N}) of whether or not the animal is in a maximum growth period}\cr
    \code{Mass} \tab {Body mass (in grams)}\cr
    \code{LogMass} \tab {Log (base 10) of body mass}\cr
    \code{Intake} \tab {Wet food intake (in grams/day)}\cr
    \code{LogIntake} \tab {Log (base 10) of Intake}\cr
    \code{WetFrass} \tab {Amount of frass (solid waste) produced (in grams/day)}\cr
    \code{LogWetFrass} \tab {Log (base 10) of WetFrass}\cr
    \code{DryFrass} \tab {Amount of frass, after drying, produced (in grams/day)}\cr
    \code{LogDryFrass} \tab {Log (base 10) of DryFrass}\cr
    \code{Cassim} \tab {CO2 assimilation (ingestion - excretion)}\cr
    \code{LogCassim} \tab {Log (base 10) of Cassim}\cr
    \code{Nfrass} \tab {Nitrogen in frass}\cr
    \code{LogNfrass} \tab {Log (base 10) of Nfrass}\cr
    \code{Nassim} \tab {Nitrogen assimilation (ingestion - excretion)}\cr
    \code{LogNassim} \tab {Log (base 10) of Nassim}\cr
  }
}
\details{
Student and faculty researchers at Kenyon College conducted
numerous experiments with Manduca Sexta caterpillars to study biological growth.
}
\source{
We thank Professors Harry Itagaki, Drew Kerkhoff, Chris Gillen, and Judy Holdener and their students for
sharing this data from research supported by NSF InSTaRs grant #0827208.
}

\keyword{datasets}
