\name{CavsShooting}
\alias{CavsShooting}
\docType{data}
\title{Cleveland Cavalier's Shooting (2016-2017)}
\description{
Shooting percentages for two Cav players
}
\format{
  A data frame with 1940 observations on the following 3 variables.
  \describe{
    \item{\code{Player}}{\code{Frye} or \code{Irving}}
    \item{\code{ShotType}}{\code{Two} or \code{Three}}
    \item{\code{Hit}}{\code{1}=made or \code{0}=missed}
  }
}
\details{
Shooting success on 2-point shots and 3-point shots for the 2016-17 NBA season for two Cleveland Cavalier basketball players, Kyrie Irving and Channing Frye. Each case is a shot attempt.  These data show Simpson's Paradox.
}
\source{
http://www.espn.com/nba/player/splits/_/id/6442/kyrie-irving
http://www.espn.com/nba/player/splits/_/id/2754/type/total/channing-frye
}
\keyword{datasets}
