\name{Cereal}
\alias{Cereal}
\docType{data}
\title{Nutrition Content of Breakfast Cereals}
\description{
Nutrition content for a sample of 36 different brands of breakfast cereals
}
\format{
  A data frame with 36 observations on the following 4 variables.
  \tabular{rl}{
    \code{Cereal} \tab {Brandname of cereal}\cr
    \code{Calories} \tab {Calories per serving}\cr
    \code{Sugar} \tab {Grams of sugar per serving}\cr
    \code{Fiber} \tab {Grams of fiber per serving}\cr
  }
}
\details{
Data give nutrition contents (per serving) for 36 breakfast cereals.
}
\source{
These data were collected by Patricia Benedict, Ronald Brahler, and
Kenneth Motz, who read the nutritional labels on the boxes,
 in an attempt to learn whether cereals high in fiber are
 also high in sugar and calories.
 The cereals are all of those that were sold at Russo Stop
  & Shop in University Heights, OH, in July, 1990.
}

\keyword{datasets}
