\name{ChildSpeaks}
\alias{ChildSpeaks}
\docType{data}
\title{Age at First Speaking}
\description{
Age at first speaking and aptitude test scores
}
\format{
  A data frame with 21 observations on the following 3 variables.
  \tabular{rl}{
    \code{Child} \tab {ID for each child}\cr
    \code{Age} \tab {Age at first speaking (in months)}\cr
    \code{Gesell} \tab {Gesell Aptitude Test Score}\cr
  }
}
\details{
The data are from a study about whether there is a relationship between the age at which a child first speaks
(in months) and his or her score on a Gesell Aptitude Test taken later in childhood.
}
\source{
These data were originally collected by L.M. Linde of UCLA but were first published by M.R. Mickey,
O.J. Dunn, and V. Clark, "Note on the use of stepwise regression in detecting outliers," Computers and Biomedical
Research, 1 (1967), pp. 105-111. The data have been used by several authors.
We found them in David Moore's Basic Practice of Statistics, WH Freeman (2004)
}

\keyword{datasets}
