\name{ClintonSanders}
\alias{ClintonSanders}
\docType{data}
\title{Clinton/Sanders Primary Results (2016)}
\description{
2016 US Democratic Presidential primary results
}
\format{
  A data frame with 31 observations on the following 5 variables.
  \describe{
    \item{\code{State}}{ID for primary state}
    \item{\code{Delegates}}{Percentage of delegates won by Clinton}
    \item{\code{PaperTrail}}{Was a paper trail available for votes cast? (\code{No Paper Trail} or \code{Paper Trail})}
    \item{\code{PopularVote}}{Percentage of votes won by Clinton}
    \item{\code{AfAmPercent}}{Percentage of African-Americans in the state}
  }
}
\details{
In 2016 Hillary Clinton won the Democratic nomination for U.S. President over Bernie Sanders. A paper was circulated that claimed to show evidence of election fraud based, among other things, on Clinton doing better in states that don't have a paper trail for votes cast in a primary election than she did in states that have a paper trail. Data is for the 31 states that held Democratic primaries in 2016.
}
\source{

https://docs.google.com/spreadsheets/d/1cszGOhbmHDTHH5ntaGPmeX55RgMMaoBhgqO1Wx-9TRk/ edit#gid=0

http://kff.org/other/state-indicator/distribution-by-raceethnicity/ ?currentTimeframe=0&sortModel=\%7B\%22colId\%22:\%22Location\%22,\%22sort\%22:\%22asc\%22\%7D

}
\keyword{datasets}
