\name{Clothing}
\alias{Clothing}
\docType{data}
\title{Sales for a Clothing Retailer}
\description{
Data on 60 customers at a clothing retailer
}
\format{
  A data frame with 60 observations on the following 8 variables.
  \tabular{rl}{
    \code{ID} \tab {Case ID}\cr
    \code{Amount} \tab {Net dollar amount spent by customers in their latest purchase from this retailer}\cr
    \code{Recency} \tab {Number of months since the last purchase}\cr
    \code{Freq12} \tab {Number of purchases in the last 12 months}\cr
    \code{Dollar12} \tab {Dollar amount of purchases in the last 12 months}\cr
    \code{Freq24} \tab {Number of purchases in the last 24 months}\cr
    \code{Dollar24} \tab {Dollar amount of purchases in the last 24 months}\cr
    \code{Card} \tab {1 for customers who have a private-label credit card with the retailer, 0 if not}\cr
  }
}
\details{
This dataset represents a random sample of 60 customers from a large clothing
retailer.  The manager of the store is interested in predicting how much a customer
will spend on his or her next purchase based on one or more of the available
explanatory variables.
}
\source{
Personal communication with David Cameron who completed a more extensive consulting project for the retailer.
}

\keyword{datasets}
