\name{CO2}
\alias{CO2}
\docType{data}
\title{CO2}
\description{
Daily carbon dioxide measurements - April through November 2011
}
%\useage{data(CO2)}
\format{
  A dataset with 237 observations on the following 2 variables.
  \tabular{rl}{
    \code{CO2} \tab {Carbon dioxide (CO2) level (in parts per million)}\cr
    \code{Day} \tab {Number of day in 2001 (April 1 = day 91}\cr
  }
}
\details{
Scientists at a research station in Brotjacklriegel, Germany recorded
CO2 levels, in parts per million, in the atmosphere for each day from
the start of April through November in 2001.
}
\source{
http://gaw.empa.ch/gawsis/reports.asp?StationID=-739519191
}

\keyword{datasets}
