\name{CO2Germany}
\alias{CO2Germany}
\docType{data}
\title{Daily CO2 Measurements in Germany}
\description{
Daily carbon dioxide measurements for April through November 2011
}
\format{
  A data frame with 237 observations on the following 2 variables.
  \describe{
    \item{\code{CO2}}{Carbon dioxide (CO2) level (in parts per million)}
    \item{\code{Day}}{Number of day in 2011 (April 1 = day 91)}
  }
}
\details{
Scientists at a research station in Brotjacklriegel, Germany recorded
CO2 levels, in parts per million, in the atmosphere for each day from
the start of April through November in 2011.
}
\source{
http://gaw.empa.ch/gawsis/reports.asp?StationID=-739519191
}
\keyword{datasets}
