\name{CO2Hawaii}
\alias{CO2Hawaii}
\docType{data}
\title{CO2 Readings in Hawaii}
\description{
Monthly carbon dioxide readings at Mauna Loa, Hawaii
}
\format{
  A data frame with 360 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (1988 - 2017)}
    \item{\code{Month}}{Month (1=Jan. to 12=Dec.)}
    \item{\code{CO2}}{Atmospheric carbon dioxide level (ppm)}
    \item{\code{t}}{Time interval (t=1 to 360)}
  }
}
\details{
Monthly average carbon dioxide readings (1988 - 2017) at the Mauna Loa Observatory in Hawaii.  Data collected and disseminated by ERSL (Earth System Research Laboratory) of the U.S. NOAA (National Oceanic and Atmospheric Administration.
}
\source{
Data downloaded for MOL (Mauna Loa) from the ESRL/GMD data page at https://www.esrl.noaa.gov/gmd/ccgg/trends/data.html
}
\keyword{datasets}
