\name{CO2SouthPole}
\alias{CO2SouthPole}
\docType{data}
\title{CO2 Readings at the South Pole}
\description{
Monthly carbon dioxide readings at the South Pole
}
\format{
  A data frame with 348 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (1988 - 2016)}
    \item{\code{Month}}{Month (1=Jan. to 12=Dec.)}
    \item{\code{CO2}}{Atmospheric carbon dioxide level (ppm)}
    \item{\code{t}}{Time interval (t=1 to 348)}
  }
}
\details{
Monthly average carbon dioxide readings (1988 - 2016) at the South Pole.  Data collected and disseminated by ERSL (Earth System Research Laboratory) of the U.S. NOAA (National Oceanic and Atmospheric Administration.
}
\source{
Data downloaded for SPO (South Pole) from the ESRL/GMD data page at https://www.esrl.noaa.gov/gmd/dv/data/
}
\keyword{datasets}
