\name{CountyHealth}
\alias{CountyHealth}
\docType{data}
\title{County Health Resources}
\description{
Medical facilities and doctors in a sample of counties.
}
\format{
  A data frame with 53 observations on the following 4 variables.
  \describe{
    \item{\code{County}}{County name, state}
    \item{\code{MDs}}{Number of medical doctors}
    \item{\code{Hospitals}}{Number of community hospitals}
    \item{\code{Beds}}{Number of beds in the hospitals}
  }
}
\details{
Data compiled from information provided by the American Medical Association on the availability of health care in counties in the United States. A random sample of 53 counties  was chosen from among counties with at least two community hospitals.
}
\source{
Physicians--American Medical Association, Chicago, IL, Physician Characteristics and Distribution in the U.S., annual (copyright), accessed May 17, 2006. Community hospitals--Health Forum LLC, an American Hospital Association (AHA) Company, Chicago, IL, Hospital Statistics, and unpublished data (copyright), e-mail accessed May 4, 2006 (related Internet site http://www.healthforum.com).

Other web sources: \cr
http://www.ama-assn.org/
http://www.healthforum.com/healthforum/html/data_statistics/data_statistics.html
http://www.cms.hhs.gov http://www.ssa.gov
}
\keyword{datasets}
