\name{CrabShip}
\alias{CrabShip}
\docType{data}
\title{Crab Oxygen Intake}
\description{
Oxygen intake of crabs with different noise sources
}
\format{
  A data frame with 34 observations on the following 3 variables.
  \describe{
    \item{\code{Mass}}{Oxygen intake of crabs with different noise sources}
    \item{\code{Oxygen}}{Rate of oxygen consumption  (mu moles h^-1)}
    \item{\code{Noise}}{Source of noise (\code{ambient} or \code{ship})}
  }
}
\details{
Animals that are stressed might increase their oxygen consumption. Biologists measured oxygen consumption of shore crabs that were either exposed to 7.5 minutes of ship noise or 7.5 minutes of ambient harbor noise.
}
\source{
Wale MA, Simpson SD, Radford AN. (2013) "Size-dependent physiological responses of shore crabs to single and repeated playback of ship noise", Biol Lett 9: 20121194. http://dx.doi.org/10.1098/rsbl.2012.1194
}
\keyword{datasets}
