\name{CrackerFiber}
\alias{CrackerFiber}
\docType{data}
\title{Effects of Cracker Fiber on Digested Calories}
\description{
Digested calories with different types of fiber in crackers
}
\format{
  A data frame with 48 observations on the following 3 variables.
  \tabular{rl}{
    \code{Subj} \tab {ID for the subject}\cr
    \code{Fiber} \tab {Type of fiber: \code{bran}, \code{combo}, \code{control}, or \code{gum}}\cr
    \code{Calories} \tab {Digested calories}\cr
  }
}
\details{
Twelve female subjects were fed a controlled diet, with crackers
before every meal. There were four different kinds of crackers: control, bran fiber, gum fiber, and
a combination of both bran and gum fiber. Over the course of the study, each subject ate all four
kinds of crackers, one kind at a time, for a stretch of several days. The order was randomized. The
response is the number of digested calories, measured as the difference between calories eaten and
calories passed through the system.
}
\source{
Subset of the data at http://lib.stat.cmu.edu/DASL/Datafiles/Fiber.html.
}

\keyword{datasets}
