\name{Cuckoo}
\alias{Cuckoo}
\docType{data}
\title{Measurements of Cuckoo Eggs}
\description{
Lengths of cuckoo eggs laid in other birds' nests
}
\format{
  A data frame with 120 observations on the following 2 variables.
  \tabular{rl}{
    \code{Bird} \tab {Type of bird nest: \code{mdw_pipit} (meadow pipit), \code{tree_pipit},}\cr
    \tab {\code{hedge_sparrow}, \code{robin}, \code{wagtail}, or \code{wren}}\cr
    \code{Length} \tab {Cuckoo egg length (in mm)}\cr
  }
}
\details{
Cuckoos are knows to lay their eggs in the nests of other (host) birds.
The eggs are then adopted and hatched by the host birds.
The data give the lengths of cuckoo eggs found in nests of various other bird species.
}
\source{
Downloaded from DASL at http://lib.stat.cmu.edu/DASL/Datafiles/cuckoodat.html
}
\references{
"The Egg of Cuculus Canorus. An Enquiry into the Dimensions of the Cuckoo's Egg and the Relation of the
Variations to the Size of the Eggs of the Foster-Parent, with Notes on Coloration", by Oswald H. Latter, Biometrika,
Vol. 1, No. 2 (Jan., 1902), pp. 164-176.
}

\keyword{datasets}
