\name{Day1Survey}
\alias{Day1Survey}
\docType{data}
\title{First Day Survey of Statistics Students}
\description{
Data from a first day class survey in an introductory statistics course
}
\format{
  A data frame with 43 observations on the following 13 variables.
  \tabular{rl}{
    \code{Section} \tab {Section: \code{1} or \code{2}}\cr
    \code{Class} \tab {Year in school: \code{Freshman},  \code{Sophomore}, \code{Junior}, or \code{Senior}}\cr
    \code{Sex} \tab {\code{F}=female or  \code{M}=male}\cr
    \code{Distance} \tab {Distance (in miles) to get to campus}\cr
    \code{Height} \tab {Height (in inches)}\cr
    \code{Handedness} \tab { \code{Left}, \code{Right}, or \code{Ambidextrous} }\cr
    \code{Coins} \tab {Value of coins student has (in class)}\cr
    \code{WhiteString} \tab {Estimated length of a white string (in inches)}\cr
    \code{BlackString} \tab {Estimated length of a black string (in inches)}\cr
    \code{Reading} \tab {Expected amount of reading during the semester (pages/week)}\cr
    \code{TV} \tab {Hours of TV watched per week}\cr
    \code{Pulse} \tab {Resting pulse rate (beats per minute)}\cr
    \code{Texting} \tab {Number of text messages in past 24 hours}\cr
  }
}
\details{
An instructor at a small liberal arts college distributed a data survey on the first day of class. The data for two
different sections of the course are given in this dataset.
}
\source{
Student survey in an introductory statistics class.
}

\keyword{datasets}
