\name{DiabeticDogs}
\alias{DiabeticDogs}
\docType{data}
\title{Lactic Acid Turnover in Dogs}
\description{
The rate of lactic acid turnover was measured by two methods for normal and diabetic dogs.
}
\format{
  A data frame with 20 observations on the following 4 variables.
  \describe{
    \item{\code{Dog}}{Code for individual dogs (\code{d1} through \code{d10})}
    \item{\code{Method}}{Tracer method to measure response (\code{infuse} or \code{inject})}
    \item{\code{Operation}}{Pancreas removed to make the dog diabetic? (\code{no} or \code{yes})}
    \item{\code{Response}}{Rate for biochemical turnover of lactic acid}
  }
}
\details{
Five dogs had their pancreas removed to make them diabetic (Operation=yes), the other five were normal (Operation=no). The rate of turnover of lactic acid was measured for each dog by two methods, infusion and injection.
}
\source{
Forbath, N., A. B. Kenshole, and G. Hetenyi, Jr. (1967),"Turnover lactic acid in normal and diabetic dogs calculated by two tracer methods," Am. J. Physiol. v. 212, pp.1179 - 1183.
}
\keyword{datasets}
