\name{Diamonds}
\alias{Diamonds}
\docType{data}
\title{Characteristics of a Sample of Diamonds}
\description{
Price and characteristics for a sample of 351 diamonds
}
\format{
  A data frame with 351 observations on the following 6 variables.
  \tabular{rl}{
    \code{Carat} \tab {Size of the diamond (in carats)}\cr
    \code{Color} \tab {Coded as \code{D} (most white/bright) through  \code{J}}\cr
    \code{Clarity} \tab {Coded as \code{IF}, \code{VVS1}, \code{VVS2}, \code{VS1}, \code{VS2}, \code{SI1}, \code{SI2}, or \code{SI3} }\cr
    \code{Depth} \tab {Depth (as a percentage of diameter)}\cr
    \code{PricePerCt} \tab {Price per carat}\cr
    \code{TotalPrice} \tab {Price for the diamond (in dollars)}\cr
  }
}
\details{
Data for a sample of diamonds. The clarity of the diamonds ranges from IF (internally flawless)
through VVS1 (very,very slightly included), VS1 (very slightly included), to SI3 (slightly included) in the order listed above.
}
\source{
Diamond data obtained from AwesomeGems.com on July 28, 2005.
}

\keyword{datasets}
