\name{Diamonds2}
\alias{Diamonds2}
\docType{data}
\title{Characteristics of a Subset of the Diamond Sample}
\description{
A subset of 307 cases with the most frequent colors from the Diamonds data
}
\format{
  A data frame with 307 observations on the following 6 variables.
  \tabular{rl}{
    \code{Carat} \tab {Size of the diamond (in carats)}\cr
    \code{Color} \tab {Coded as \code{D} (most white/bright) through  \code{G}}\cr
    \code{Clarity} \tab {Coded as \code{IF}, \code{VVS1}, \code{VVS2}, \code{VS1}, \code{VS2}, \code{SI1}, \code{SI2}, or \code{SI3} }\cr
    \code{Depth} \tab {Depth (as a percentage of diameter)}\cr
    \code{PricePerCt} \tab {Price per carat}\cr
    \code{TotalPrice} \tab {Price for the diamond (in dollars)}\cr
  }
}
\details{
A subset of the Diamonds data, containing only those with most frequent colors D, E, F, and G.
The clarity of the diamonds ranges from IF (internally flawless)
through VVS1 (very,very slightly included), VS1 (very slightly included),
to SI3 (slightly included) in the order listed above.
}
\source{
Diamond data obtained from AwesomeGems.com on July 28, 2005.
}

\keyword{datasets}
