\name{Dinosaurs}
\alias{Dinosaurs}
\docType{data}
\title{Iridium Levels in Rock Layers to Investigate Dinosaur Extinction}
\description{
Iridium levels in prehistoric rock layers
}
\format{
  A data frame with 28 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{Sample identifier}
    \item{\code{Source}}{Type of rock (\code{Limestone} \code{Shale})}
    \item{\code{Depth}}{Depth of the sample (in meters)}
    \item{\code{Iridium}}{Iridium concentration (ppb)}
  }
}
\details{
The question of interest is whether a volcanic eruption or asteroid strike had created a dust cloud that led to extinction of most dinosaurs.  Rock samples taken in Gubbio, Italy were measured for the concentration of iridium (a rare metal which is more common in asteroids).  The deeper the sample, the older the rocks are.  A sudden increase in iridium at some point in time would lend support for the asteroid hypothesis.
}
\source{
Ramsey, Fred L. and Daniel W. Schafer (2002). The Statistical Sleuth, 2nd ed., Pacific Grove, CA, Duxbury, pp.405-407.
}
\keyword{datasets}
