\name{Election08}
\alias{Election08}
\docType{data}
\title{2008 U.S. Presidential Election}
\description{
State-by-state information from the 2008 U.S. presidential election
}
\format{
  A dataframe with 51 observations on the following 7 variables.
  \tabular{rl}{
    \code{State} \tab {Name of the state}\cr
    \code{Abr} \tab {Abbreviation for the state}\cr
    \code{Income} \tab {Per capita income in the state as of 2007 (in dollars)}\cr
    \code{HS} \tab {Percentage of adults with at least a high school education}\cr
    \code{BA} \tab {Percentage of adults with at least a college education}\cr
    \code{Dem.Rep} \tab {Difference in \%Democrat and \%Republican (according to 2008 Gallup survey)}\cr
    \code{ObamaWin} \tab {\code{1}= Obama (Democrat) wins state in 2008 or \code{0}=McCain (Republican) wins}\cr
  }
}
\details{
This dataset contains information from all 50 states and the District of Columbia for the 2008 U.S. presidential election.
}
\source{
State income data from: Census Bureau Table 659. Personal Income Per Capita (in 2007)\cr
High school data from: U.S. Census Bureau, 1990 Census of Population, \cr
http://nces.ed.gov/programs/digest/d08/tables/dt08_011.asp \cr
College data from: Census Bureau Table 225. Educational Attainment by State (in 2007)\cr
\% Democrat and \%Republican: \cr
http://www.gallup.com/poll/114016/state-states-political-party-affiliation.aspx#1
}

\keyword{datasets}
