\name{ElephantsMF}
\alias{ElephantsMF}
\docType{data}
\title{Measurements of African Elephants}
\description{
Age and height of African elephants
}
\format{
  A data frame with 288 observations on the following 3 variables.
  \describe{
    \item{\code{Age}}{Age (in years)}
    \item{\code{Height}}{Shoulder height (in cm)}
    \item{\code{Sex}}{\code{F}=female or \code{M}=male}
  }
}
\details{
Data on 288 African elephants that lived through droughts in the first two years of life.
}
\source{
Data are from Phyllis Lee, Stirling University, and are related to Lee, P., et al. (2013), "Enduring consequences of early experiences: 40-year effects on survival and success among African elephants (Loxodonta Africana)," Biology Letters, 9: 20130011.

}
\keyword{datasets}
