% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplogitplot1.r
\name{emplogitplot1}
\alias{emplogitplot1}
\title{Empirical logit plot for one quantitative variable}
\usage{
emplogitplot1(formula, data = NULL, ngroups = 3, breaks = NULL,
  yes = NULL, padj = TRUE, out = FALSE, showplot = TRUE,
  showline = TRUE, ylab = "Log(Odds)", xlab = NULL,
  dotcol = "black", linecol = "blue", pch = 16, main = "",
  ylim = NULL, xlim = NULL, lty = 1, lwd = 1, cex = 1)
}
\arguments{
\item{formula}{A formula of the form (binary) Response~Predictor}

\item{data}{A dataframe}

\item{ngroups}{Number of groups to use (not needed if breaks is used), ngroups="all" uses all unique values}

\item{breaks}{A vector of endpoints for the bins (not needed if ngroups is used)}

\item{yes}{Set a value for the response to be counted for proportions (optional)}

\item{padj}{Should proportions be adjusted to avoid zero and one? (default is TRUE)}

\item{out}{Should the function return a dataframe with group information? (default is FALSE)}

\item{showplot}{Show the plot? default is TRUE}

\item{showline}{Show the regression line? default is TRUE}

\item{dotcol}{Color for the dots (default is "black")}

\item{linecol}{Color for the line (default is "blue")}
}
\value{
A dataframe with group information (if out=TRUE)
}
\description{
This function produces an empirical logit plot for a binary response variable and a
single quantitative predictor variable.
}
\details{

}
\examples{
data(MedGPA)
emplogitplot1(Acceptance~GPA,data=MedGPA)

data(Putts1)
}
