\name{Ethanol}
\alias{Ethanol}
\docType{data}
\title{Ethanol}
\description{
Ethanol
}
%\useage{data(Ethanol)}
\format{
  A dataset with 16 observations on the following 3 variables.
  \tabular{rl}{
    \code{Sugar} \tab {Type of sugar:  \code{Galactose} or  \code{Glucose}}\cr
    \code{O2Conc} \tab {Oxygen concentration}\cr
    \code{Ethanol} \tab {Ethanol concentration}\cr
  }
}
\details{
Many biochemical reactions are slowed or prevented by the presence of oxygen. For
example, there are two simple forms of fermentation, one which converts each molecule of sugar
to two molecules of lactic acid, and a second which converts each molecule of sugar to one each of
lactic acid, ethanol, and carbon dioxide. This experiment was designed to compare the inhibiting effect of oxygen on the
metabolism of two different sugars, glucose and galactose, by Streptococcus bacteria. In this case
there were four levels of oxygen that were applied to the two kinds of sugar.

Renamed to SugarEthanol in second edition.
}
\source{
Data are found in
Statistics: The Exploration and Analysis of Data by Jay Devore and Roxy Peck (1960). St. Paul, MN: West.
}
\references{
The original article is "Effects of oxygen concentration on pyruvate formatelyase in situ and sugar metabolism
of Streptocucoccus mutans and Streptococcus samguis," Infection and Immunity, pp. 129-134.
}

\keyword{datasets}
