\name{Faces}
\alias{Faces}
\docType{data}
\title{Facial Attractiveness of Men}
\description{
Grip strength, attractiveness, and shoulder-hip ratio for men
}
\format{
  A data frame with 38 observations on the following 5 variables.
  \describe{
    \item{\code{MaxGripStrength}}{Measurement of strength of hand grip}
    \item{\code{SHR}}{Shoulder to hip ratio}
    \item{\code{Partners}}{Number of sexual partners (lifetime)}
    \item{\code{Attractive}}{Attractiveness rating}
    \item{\code{AgeFirstSex}}{Age of first sex}
  }
}
\details{
Facial attractiveness of several men was rated by female college students. Maximum grip strength was also measured, along with shoulder to hip ratio, age of first sex, and number of sex partners.
}
\source{
Shoup, M. L. and Gallup, G.G., Jr. (2008), "Men's Faces Convey Information about Their Bodies and Their Behavior: What You See is What You Get," Evolutionary Psychology, 6(3): 469-479.
}
\keyword{datasets}
