\name{FantasyBaseball}
\alias{FantasyBaseball}
\docType{data}
\title{FantasyBaseball}
\description{
Draft selection times for a fantasy baseball league
}
%\useage{data(FantasyBaseball)}
\format{
  A dataset with 24 observations on the following 9 variables.
  \tabular{rl}{
    \code{Round} \tab {Round of the draft (1 to 24)}\cr
    \code{DJ} \tab {Draft time (in seconds) for D.J.}\cr
    \code{AR} \tab {Draft time (in seconds) for A.R.}\cr
    \code{BK} \tab {Draft time (in seconds) for B.K.}\cr
    \code{JW} \tab {Draft time (in seconds) for J.W.}\cr
    \code{TS} \tab {Draft time (in seconds) for T.S.}\cr
    \code{RL} \tab {Draft time (in seconds) for R.L.}\cr
    \code{DR} \tab {Draft time (in seconds) for D.R.}\cr
    \code{MF} \tab {Draft time (in seconds) for M.F.}\cr
  }
}
\details{
Time (in seconds) for participants in a draft for a fantasy baseball league to make a selectoion at each round. 
}
\source{
Mathamatical Science Baseball League historical records (online). 
}

\keyword{datasets}
