\name{FatRats}
\alias{FatRats}
\docType{data}
\title{Diet and Weight of Rats}
\description{
Experiment on effects of diets on weight gain of rats
}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{\code{Gain}}{Weight gain (in grams per week)}
    \item{\code{Protein}}{Level of protein (\code{Hi} or \code{Lo})}
    \item{\code{Source}}{Source of protein (\code{Beef}, \code{Cereal}, or \code{Pork})}
  }
}
\details{
Data from this experiment compared weight gain for 60 baby rats that were fed different diets.  Half of the rats had low-protein diets (Lo) and the rest had high-protein (Hi).  The source of protein was either beef, cereal, or pork.
}
\source{
C. P. Wilsie, Iowa State College Agricultural Station (1944) via Snedecor and Cochran
}
\keyword{datasets}
