\name{Fertility}
\alias{Fertility}
\docType{data}
\title{Fertility Data for Women Having Trouble Getting Pregnant}
\description{
Fertility measurements for a sample of women who have difficulty getting pregnant
}
\format{
  A data frame with 333 observations on the following 10 variables.
  \tabular{rl}{
    \code{Age} \tab {Age (in years)}\cr
    \code{LowAFC} \tab {Smallest antral follicle count}\cr
    \code{MeanAFC} \tab {Average antral follicle count}\cr
    \code{FSH} \tab {Maximum follicle stimulating hormone level}\cr
    \code{E2} \tab {Fertility level}\cr
    \code{MaxE2} \tab {Maximum fertility level}\cr
    \code{MaxDailyGn} \tab {Maximum daily gonadotropin level}\cr
    \code{TotalGn} \tab {Total gonadotropin level}\cr
    \code{Oocytes} \tab {Number of egg cells}\cr
    \code{Embryos} \tab {Number of embryos}\cr
  }
}
\details{
A medical doctor and her team of researchers collected a
variety of data on women who were having trouble getting pregnant.
A key method for assessing fertility is a count of antral follicles (LowAFC or MeanAFC)
that can be performed with noninvasive ultrasound.
Researchers are interested in how the other variables are related to these counts.
}
\source{
We thank Dr. Priya Maseelall and her research team for sharing these data.
}

\keyword{datasets}
