\name{Fingers}
\alias{Fingers}
\docType{data}
\title{Fingers}
\description{
Fingers
}
%\useage{data(Fingers)}
\format{
  A dataset with 12 observations on the following 3 variables.
  \tabular{rl}{
    \code{Subject} \tab {\code{I}, \code{II}, \code{III}, or \code{IV}}\cr
    \code{Drug} \tab {\code{Caffeine}, \code{Placebo}, or \code{Theobromine}}\cr
    \code{TapRate} \tab {Finger taps in a fixed time interval}\cr
  }
}
\details{
Scientists Scott and Chen, published
research that compared the effects of caffeine with those of theobromine (a similar chemical
found in chocolate) and with those of a placebo. Their experiment used four human subjects, and
took place over several days. Each day each subject swallowed a tablet containing one of caffeine,
theobromine, or the placebo. Two hours later they were timed while tapping a finger in a specified
manner (that they had practiced earlier, to control for learning effects). The response is the number
of taps in a fixed time interval.

Renamed FranticFingers in second edition.
}
\source{
The data was found in Statistics in Biology, Vol. 1, by C. I. Bliss (1967), New York: McGraw Hill.
}
\references{
1The original article is "Comparison of the action of 1-ethyl theobromine and caffeine in animals and man," by
C. C. Scott and K. K. Chen, Journal of Pharmacological Experimental Therapy, v. 82 (1944), pp 89-97.
}

\keyword{datasets}
