\name{FishEggs}
\alias{FishEggs}
\docType{data}
\title{Fertility of Fish Eggs}
\description{
Fertility measurement for eggs from a sample of 35 lake trout
}
\format{
  A data frame with 35 observations on the following 4 variables.
  \tabular{rl}{
    \code{Age} \tab {Age of the fish (in years)}\cr
    \code{PctDM} \tab {Percentage of the total egg material that is solid}\cr
    \code{Month} \tab {Month fish was caught:  \code{Sep}=September or  \code{Nov}=November}\cr
    \code{Sept} \tab {Indicator with \code{1}=September or \code{0}=November}\cr
  }
}
\details{
Researchers collected samples of female lake trout from Lake Ontario in September
and November of 2002 through 2004. A goal of the study was to investigate the fertility of
fish that had been stocked in the lake. One measure of the viability of fish eggs is percent dry mass (PctDM) which reflects the energy potential stored in the eggs by recording the percentage
of the total egg material that is solid. Values of the PctDM for a sample of 35 lake trout (14 in
September and 21 in November) are given in this dataset along with the age (in years) of the fish.
}
\source{
Lantry, OGorman, and Machut (2008) "Maternal Characteristics versus Egg Size and Energy Density," Journal
of Great Lakes Research 34(4): 661-674.
}

\keyword{datasets}
