\name{Fitch}
\alias{Fitch}
\docType{data}
\title{Body Measurements of Mammal Species}
\description{
Body measurements for a sample of 28 mammal species from a Fitch paper on acoustic allometry
}
\format{
  A data frame with 28 observations on the following 5 variables.
  \describe{
    \item{\code{Species}}{species of mammal}
    \item{\code{Order}}{Order (\code{Carnivora} or \code{Primates})}
    \item{\code{Wt}}{Body weight (in kg)}
    \item{\code{Skull}}{Skull length (in cm)}
    \item{\code{Palate}}{Palate length (in cm)}
  }
}
\details{
Data on mammal species from a Zoology paper about acoustic allometry by W. Tecumseh Fitch.
}
\source{
Fitch, W. Tecumseh (2000), "Skull dimensions in relation to body size in nonhuman mammals: The causal bases for acoustic allometry," Zoology, 103, 40-58.
}
\keyword{datasets}
