\name{FloridaDP}
\alias{FloridaDP}
\docType{data}
\title{Florida Death Penalty Cases}
\description{
Florida death penalty cases by race of defendant and victim
}
\format{
  A data frame with 326 observations on the following 4 variables.
  \describe{
    \item{\code{Penalty}}{Was death penalty given? (\code{No} or \code{Yes})}
    \item{\code{Defendant}}{Race of the defendant (\code{Black} or \code{White})}
    \item{\code{White.Victim}}{Was the victim white? (\code{1}=yes or \code{0}=no)}
    \item{\code{Black.Victim}}{Was the victim black? (\code{1}=yes or \code{0}=no)}
  }
}
\details{
Mike Radelet's data on imposition of the death penalty for murderers in Florida broken down by race of the victim and defendant.
}
\source{
Radelet, M. (1981), "Racial Characteristics and Imposition of the Death Penalty," American Sociological Review, 46, 918-927.
}
\keyword{datasets}
