\name{Fluorescence}
\alias{Fluorescence}
\docType{data}
\title{Measuring Calcium Binding to Proteins}
\description{
Data from an experiment on calcium binding to proteins
}
\format{
  A data frame with 51 observations on the following 2 variables.
  \tabular{rl}{
    \code{Calcium} \tab {Log of free calcium concentration}\cr
    \code{ProteinProp} \tab {Proportion of protein bound to calcium}\cr
  }
}
\details{
Suzanne Rohrback used a novel approach in a series of experiments to examine calcium binding proteins.
}
\source{
Thanks to Suzanne Rohrback for providing these data from her honors experiments at Kenyon College.
}

\keyword{datasets}
