\name{FruitFlies}
\alias{FruitFlies}
\docType{data}
\title{Fruit Fly Sexual Activity and Longevity}
\description{
Sexual activity and lifetimes of fruit flies
}
\format{
  A data frame with 125 observations on the following 7 variables.
  \tabular{rl}{
    \code{ID} \tab {a numeric vector}\cr
    \code{Partners} \tab {Number of female partners: 0, 1, or 8}\cr
    \code{Type} \tab {\code{0}=pregnant, \code{1}=virgin, \code{9}=none}\cr
    \code{Longevity} \tab {Lifespan (in days)}\cr
    \code{Thorax} \tab {Length of thorax (in mm)}\cr
    \code{Sleep} \tab {Percent of day sleeping}\cr
   \code{Treatment} \tab {\code{1 pregnant}, \code{1 virgin}, \code{8 pregnant}, \code{8 virgin}, or \code{none}}\cr
  }
}
\details{
Hanley and Shapiro (1994) report on a
study conducted by Partridge and Farquhar (1981) about the sexual behavior of fruit flies. It was
already known that increased reproduction leads to shorter life spans for female fruit flies. But the
question remained whether an increase in sexual activity would also reduce the life spans of male
fruit flies. The researchers designed an experiment to answer this question. They had a total of
125 male fruit flies to use and they randomly assigned each of the 125 to one of the following five
groups.
}
\source{
The data are given as part of the data archive on the Journal of Statistics Education website
and can be found on the page \cr
http://www.amstat.org/publications/jse/jse_data_archive.htm.
}
\references{
Hanley and Shapiro, (1994) "Sexual Activity and the Lifespan of Male Fruitflies: A Dataset That Gets Attention," Journal of Statistics Education
v.2, n.1\cr
http://www.amstat.org/publications/jse/v2n1/datasets.hanley.html
}

\keyword{datasets}
