\name{Goldenrod}
\alias{Goldenrod}
\docType{data}
\title{Goldenrod Galls}
\description{
Measurements for a sample of goldenrod galls
}
\format{
  A data frame with 1055 observations on the following 9 variables.
  \tabular{rl}{
    \code{Gdiam03} \tab {Gall diameter in 2003 (in mm)}\cr
    \code{Stdiam03} \tab {Stem diameter in 2003 (in mm)}\cr
    \code{Wall03} \tab {Wall thickness in 2003 (in mm)}\cr
    \code{Fate03} \tab {\code{b}=beetle present \code{e}=early death \code{f}=living fly larva \code{g}=living wasp  \code{o}=pupal case \code{u}=unknown}\cr
    \code{Gdiam04} \tab {Gall diameter in 2004 (in mm)}\cr
    \code{Stdiam04} \tab {Stem diameter in 2004 (in mm)}\cr
    \code{Wall04} \tab {Wall thickness in 2003 (in mm)}\cr
    \code{Fate04} \tab {\code{b}=beetle present \code{e}=early death \code{f}=living fly larva \code{g}=living wasp  \code{o}=pupal case \code{u}=unknown}\cr
    \code{Fly04} \tab {Fly in 2004? \code{n} or \code{y}}\cr
  }
}
\details{
Biology students collected measurements on goldenrod galls at the
Brown Family Environmental Center at Kenyon College.
}
\source{
Thanks to the Kenyon College Department of Biology for sharing these data.
}

\keyword{datasets}
