\name{Grocery}
\alias{Grocery}
\docType{data}
\title{Grocery Sales and Discounts}
\description{
Grocery store sales with different discounts
}
\format{
  A data frame with 36 observations on the following 5 variables.
  \tabular{rl}{
    \code{Discount} \tab {Amount of discount: \code{5.00\%}, \code{10.00\%}, or \code{15.00\%}}\cr
    \code{Store} \tab {Store number (1-12)}\cr
    \code{Display} \tab {\code{Featured End of Aisl}, \code{Featured Middle of A}, or \code{Not Featured}}\cr
    \code{Sales} \tab {Number sold during one week}\cr
    \code{Price} \tab {Wholesale price (in dollars)}\cr
  }
}
\details{
Grocery stores and product manufacturers are always interested in how well the products on the
store shelves sell. An experiment was designed to test whether the amount of discount given on
products affected the amount of sales of that product. There were three levels of discount, 5\%,
10\%, and 15\%, and sales were held for a week. The total number of products sold during the week
of the sale was recorded. The researchers also recorded the wholesale price of the item put on sale.
}
\source{
These data are not real, though they are simulated to approximate an actual study. The data come from John
Grego, Director of the Stat Lab at University of South Carolina.
}

\keyword{datasets}
