\name{Gunnels}
\alias{Gunnels}
\docType{data}
\title{Are Gunnels Present at Shoreline?}
\description{
Presence/absence of gunnels (eels) at shoreline quadrats
}
\format{
  A data frame with 1592 observations on the following 10 variables.
  \tabular{rl}{
    \code{Gunnel} \tab {1= gunnel present in the quadrat or 0=gunnel absent}\cr
    \code{Time} \tab {Minutes after midnight}\cr
    \code{Fromlow} \tab {Time in minutes from low tide}\cr
    \code{Slope} \tab {Slope (to nearest 10 degrees) perpendicular to waterline}\cr
    \code{Rw} \tab {Percentage cover in quadrat of rockweed/algae/plants}\cr
    \code{Amphiso} \tab {Density of crustacean food: 0=none to 4=high}\cr
    \code{Subst} \tab {Substratum: 1=solid rock, 2=rocky cobbles, 3=mixed pebbles/sand, 4=fine sand,}\cr
    \tab {5=mud, 6=mixed mud/shell detritus, 7=cobbles on solid rock, 8=cobbles on mixed pebbles/sand,}\cr
    \tab {9=cobbles on fine sand, 10=cobbles on mud, 11=cobbles on mixed mud/shell detritus,}\cr
    \tab {12=cobbles on shell detritus, 13=shell detritus }\cr
    \code{Pool} \tab {Standing water deep?  1=yes or 2=no}\cr
    \code{Water} \tab {Standing water in the quadrat?  1=yes or 2=no}\cr
    \code{Cobble} \tab {Rocky cobbles? 1=yes or 2=no }\cr
  }
}
\details{
This dataset comes from a study on the habitat preferences of a species
of eel, called a gunnel. Biologist Jake Shorty sampled quadrats along a coastline and recorded whether or not
the species was found in the quadrat.
}
\source{
Thanks to Jake Shorty, Bowdoin biology student, for this dataset.
}

\keyword{datasets}
