\name{Hawks}
\alias{Hawks}
\docType{data}
\title{Measurements on Three Hawk Species}
\description{
Data for a samples of hawks from three different species
}
\format{
  A data frame with 908 observations on the following 19 variables.
  \tabular{rl}{
    \code{Month} \tab {\code{8}=September to \code{12}=December}\cr
    \code{Day} \tab {Date in the month}\cr
    \code{Year} \tab {Year: 1992-2003}\cr
    \code{CaptureTime} \tab {Time of capture (HH:MM)}\cr
    \code{ReleaseTime} \tab {Time of release (HH:MM)}\cr
    \code{BandNumber} \tab {ID band code}\cr
    \code{Species} \tab {\code{CH}=Cooper's, \code{RT}=Red-tailed,  \code{SS}=Sharp-Shinned}\cr
    \code{Age} \tab {\code{A}=Adult or \code{I}=Imature}\cr
    \code{Sex} \tab {\code{F}=Female or \code{M}=Male}\cr
    \code{Wing} \tab {Length (in mm) of primary wing feather from tip to wrist it attaches to}\cr
    \code{Weight} \tab {Body weight (in gm)}\cr
    \code{Culmen} \tab {Length (in mm) of the upper bill from the tip to where it bumps into the fleshy part of the bird}\cr
    \code{Hallux} \tab {Length (in mm) of the killing talon}\cr
    \code{Tail} \tab {Measurement (in mm) related to the length of the tail (invented at the MacBride Raptor Center)}\cr
    \code{StandardTail} \tab {Standard measurement of tail length (in mm)}\cr
    \code{Tarsus} \tab {Length of the basic foot bone (in mm)}\cr
    \code{WingPitFat} \tab {Amount of fat in the wing pit}\cr
    \code{KeelFat} \tab {Amount of fat on the breastbone (measured by feel}\cr
    \code{Crop} \tab {Amount of material in the crop, coded from \code{1}=full to \code{0}=empty}\cr
  }
}
\details{
Students and faculty at Cornell College in Mount Vernon, Iowa, collected  data over many years
at the hawk blind at Lake MacBride near Iowa City, Iowa. The data set that we are analyzing here
 is a subset of the original data set, using only those species for which there were more than 10 observations.
 Data were collected on random samples of three different species of hawks:
 Red-tailed, Sharp-shinned, and Cooper's hawks.
}
\source{
Many thanks to the late Professor Bob Black at Cornell College for sharing these data with us.
}

\keyword{datasets}
