\name{HawkTail}
\alias{HawkTail}
\docType{data}
\title{HawkTail}
\description{
Tail lengths for two hawk species
}
%\useage{data(HawkTail)}
\format{
  A dataset with 838 observations on the following 2 variables.
  \tabular{rl}{
    \code{Species} \tab {\code{RT}=Red-tailed, \code{SS}=Sharp-shinned}\cr
    \code{Tail} \tab {Length of tail (in mm)}\cr
  }
}
\details{
Tail lengths measured for a sample of 838 hawks observed in Mount Vernon, Iowa. 
Note: HawkTail2 has similar data in unstacked format for three species of hawks.
}
\source{
Observations by students and faculty at Cornell College.
}

\keyword{datasets}
