\name{HawkTail2}
\alias{HawkTail2}
\docType{data}
\title{HawkTail2}
\description{
Tail lengths for two hawk species
}
%\useage{data(HawkTail2)}
\format{
  A dataset with observations on the following 2 variables.
  \tabular{rl}{
    \code{Tail_RT} \tab {Tail length (in mm) for a sample of Red-tailed hawks}\cr
    \code{Tail_SS} \tab {Tail length (in mm) for a sample of Sharp-shinned hawks}\cr
  }
}
\details{
Tail lengths measured for a sample of hawks observed in Mount Vernon, Iowa. 
Note: HawkTail has similar data in stacked format for just the Red-tailed and Sharp-shinned hawks. The Hawks dataset jhas more varialbes and a third species (Cooper's Hawk)
}
\source{
Observations by students and faculty at Cornell College.
}

\keyword{datasets}
