\name{HeatingOil}
\alias{HeatingOil}
\docType{data}
\title{Heating Oil Consumption}
\description{
Monthly US residential consumption of fuel oil (1983-2016)
}
\format{
  A data frame with 408 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (1983 to 2016)}
    \item{\code{Month}}{Month (1=Jan through 12=Dec)}
    \item{\code{t}}{Time index (1 to 408)}
    \item{\code{FuelOil}}{Residential consumption of fuel oil (in 1,000 barrels/day)}
  }
}
\details{
U.S. residential consumption of distillate fuel oil each month from January 1983 through December 2016.
}
\source{
U.S. Energy Information Administration website, https://www.eia.gov/totalenergy/data/monthly/index.php
}
\keyword{datasets}
