\name{HighPeaks}
\alias{HighPeaks}
\docType{data}
\title{Characteristics of Adirondack Hiking Trails}
\description{
Data on hiking trails for each of the 46 "High Peaks" in the Adirondack mountains
}
\format{
  A data frame with 46 observations on the following 6 variables.
  \tabular{rl}{
    \code{Peak} \tab {Name of the mountain}\cr
    \code{Elevation} \tab {Elevation at the highest point (in feet)}\cr
    \code{Difficulty} \tab {Rating of difficulty of the hike: \code{1} (easy) to \code{7} (most difficult) }\cr
    \code{Ascent} \tab {Vertical ascent (in feet)}\cr
    \code{Length} \tab {Length of hike (in miles)}\cr
    \code{Time} \tab {Expected trip time (in hours)}\cr
  }
}
\details{
Forty-six mountains in the Adirondacks of upstate New York are
known as the High Peaks with elevations near or above 4000 feet (although modern measurements
show a couple of the peaks are actually slightly under 4000 feet). A goal for hikers in the region is
to become a "46er" by scaling each of these peaks.
This dataset gives information about the hiking trails up each of these peaks.
}
\source{
High Peaks data avaialble at http://www.adirondack.net/tour/hike/highpeaks.cfm.
Thanks to Jessica Chapman at St. Lawrence University for recommending this dataset.
}

\keyword{datasets}
