\name{HorsePrices}
\alias{HorsePrices}
\docType{data}
\title{Prices of Horses}
\description{
Price and related characteristics of horses listed for sale on the internet
}
\format{
  A data frame with 50 observations on the following 5 variables.
  \tabular{rl}{
    \code{HorseID} \tab {ID code for each horse}\cr
    \code{Price} \tab {Price (in dollars)}\cr
    \code{Age} \tab {Age of the horse (in years)}\cr
    \code{Height} \tab {Height of the horse (in hands)}\cr
    \code{Sex} \tab {\code{f}=female \code{m}=male}\cr
  }
}
\details{
Undergraduate students at Cal Poly collected data on prices of 50 horses
advertised for sale on the internet. Predictor variables of price include the age and height of the horse (in
hands), as well as its sex.
}
\source{
Cal Poly students using a horse sale website.
}

\keyword{datasets}
