\name{Houses}
\alias{Houses}
\docType{data}
\title{Houses}
\description{
Houses
}
%\useage{data(Houses)}
\format{
  A dataset with 20 observations on the following 3 variables.
  \tabular{rl}{
    \code{Price} \tab {Selling price (in dollars)}\cr
    \code{Size} \tab {Size of the house (in square feet)}\cr
    \code{Lot} \tab {Area of the house's lot (in square feet)}\cr
  }
}
\details{
This dataset contains selling prices for 20 houses that were sold in 2008 in a small midwestern
town. The file also contains data on the size of each house (in square feet) and the size of the lot
(in square feet) that the house is on.

Updated to HousesNY in second edition.
}
\source{
Data collected from zillow.com in June 2008.
}

\keyword{datasets}
