\name{HousesNY}
\alias{HousesNY}
\docType{data}
\title{House Prices in Rural NY}
\description{
House prices for a sample of houses in Canton NY
}
\format{
  A data frame with 53 observations on the following 5 variables.
  \describe{
    \item{\code{Price}}{Estimated price (in $1,000's)}
    \item{\code{Beds}}{Number of bedrooms}
    \item{\code{Baths}}{Number of bathrooms}
    \item{\code{Size}}{Floor area of the house (in 1,000 square feet)}
    \item{\code{Lot}}{Size of the lot (in acres)}
  }
}
\details{
Data scraped from Zillow.com for a sample of houses near the 13617 area code (Canton, NY a small town in upstate NY).  Houses on lots bigger than five acres (often farms) were excluded.
}
\source{
Data scraped from the Zillow.com website using tools an app at http://myslu.stlawu.edu/~clee/dataset/zillow/  (April 2017)
}
\keyword{datasets}
