\name{InfantMortality}
\alias{InfantMortality}
\docType{data}
\title{InfantMortality}
\description{
Infant mortality rates in the United States (1920-2000)
}
%\useage{data(InfantMortality)}
\format{
  A dataset with 9 observations on the following 2 variables.
  \tabular{rl}{
    \code{Mortality} \tab {Deaths witin one year of birth (per 1000 births}\cr
    \code{Year} \tab {1920-2000 (by decades)}\cr
  }
}
\details{
Infant mortality (deaths within one year of birth per 1,000 births) in the US from 1920 - 2000 (by decade).
}
\source{
CDC National Vital Statistics Reports at http://www.cdc.gov/nchs/data/nvsr/nvsr57/nvsr57 14.pdf
}

\keyword{datasets}
