\name{Inflation}
\alias{Inflation}
\docType{data}
\title{Monthly Consumer Price Index (2009-2016)}
\description{
Consumer Price Index (CPI) each month for 2009 through 2016
}
\format{
  A data frame with 96 observations on the following 5 variables.
  \describe{
    \item{\code{Month}}{Month: \code{1}=January to \code{12}=December}
    \item{\code{Year}}{Year (2009 to 2016)}
    \item{\code{CPI}}{Consumer Price Index (base=100 in 1984)}
    \item{\code{CPIPctDiff}}{Monthly percent change in CPI}
    \item{\code{t}}{Time index (1 to 96)}
  }
}
\details{
Monthly Consumer Price Index for 2009 to 2016 as produced by the Bureau of Labor Statistics (Series Id. CUUR0000SA0).  Based on prices for all items in U.S. city average for all consumers (not seasonally) Base period is 1982-1984-100.
}
\source{
Data downloaded from Bureau of Labor Statistics at\cr
https://www.bls.gov/data/
}
\keyword{datasets}
