\name{IQGuessing}
\alias{IQGuessing}
\docType{data}
\title{Guess IQ from a Photo?}
\description{
True IQ and guessed IQ (from a photo) for 40 women
}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{Age}}{Age of woman}
    \item{\code{GuessIQ}}{Guessed IQ}
    \item{\code{TrueIQ}}{Actual IQ}
  }
}
\details{
One hundred sixty raters (75 men and 85 women) took part in judging intelligence (on a 1=high to 7=low scale) based on photographs of students. The ratings were converted to z-scores and then put on an IQ scale to compare to actual measured IQ. There were photos of 80 students, 40 men and 40 women.  This data set contains data for the 40 women.
}
\source{
Kleisner K, Chvatalova V, Flegr J (2014), "Perceived Intelligence Is Associated with Measured Intelligence in Men but Not Women," PLoS ONE 9(3): e81237. doi:10.1371/journal.pone.0081237.
}
\keyword{datasets}
