\name{KeyWestWater}
\alias{KeyWestWater}
\docType{data}
\title{Key West Water Temperatures}
\description{
Hourly water temperatures from Gulf of Mexico near Key West, Florida
}
\format{
  A data frame with 6572 observations on the following 3 variables.
  \describe{
    \item{\code{DateTime}}{Date and time of reading (format \code{mm/dd/yyyy h:00})}
    \item{\code{WaterTemp}}{Water temperature (in degrees Fahrenheit)}
    \item{\code{t}}{Time index (1 to 673)}
  }
}
\details{
Hourly readings of water temperatures from a measuring device in the Gulf of Mexico near Key West, Florida.  The hourly temperatures are provided from October 3, 2016 to October 3, 2017 and were obtained from station 8724580.  A few missing values have been interpolated to provide a complete series.
}
\source{
National Oceanographic and Atmospheric Administration (2017), Key West Ocean Temperature Data, October 3, 2016 to October 3, 2017, https://www.nodc.noaa.gov, Accessed on October 4, 2017

Data were obtained by Kyle Johnston for his Senior Exercise (a capstone project).
}
\keyword{datasets}
