\name{Kids198}
\alias{Kids198}
\docType{data}
\title{Kids198}
\description{
Measurements of Children
}
%\useage{data(Kids198)}
\format{
  A dataset with 198 observations on the following 5 variables.
  \tabular{rl}{
    \code{Height} \tab {Height (in inches)}\cr
    \code{Weight} \tab {Weight (in pounds)}\cr
    \code{Age} \tab {Age (in months)}\cr
    \code{Sex} \tab {\code{0}=male or \code{1}=female}\cr
    \code{Race} \tab {\code{0}=white or \code{1}=other}\cr
  }
}
\details{
This dataset comes from a 1977 anthropometric study of body measurements for children.  
Subjects in this sample are between the ages of 8 and 18 years old, 
selected at random from the much larger dataset of the oringinal study. 
}
\source{
A sample of 198 cases from the NIST's AnthroKids dataset at http://ovrt.nist.gov/projects/anthrokids/
}

\keyword{datasets}
