\name{LeafWidth}
\alias{LeafWidth}
\docType{data}
\title{Leaf Measurements}
\description{
Measurements of Dodonaea viscosa leaves
}
\format{
  A data frame with 252 observations on the following 5 variables.
  \describe{
    \item{\code{Width}}{Average width (in mm)}
    \item{\code{Length}}{Average length  (in mm)}
    \item{\code{LWRatio}}{Length divided by Width}
    \item{\code{Area}}{Area (in sq. mm) }
    \item{\code{Year}}{Year the leaves were collected}
  }
}
\details{
Data on samples of leaves from the species Dodonaea viscosa subsp. angustissima (common name hopbush), which have been collected in a certain region of South Australia for many years.
}
\source{
Guerin, G., Wen, H., Lowe, A. (2012), "Leaf morphology shift linked to climate change," Biol. Lett., 8, doi: 10.1098/rsbl.2012.0458
}
\keyword{datasets}
