\name{Leukemia}
\alias{Leukemia}
\docType{data}
\title{Responses to Treatment for Leukemia}
\description{
Treatment results for leukemia patients
}
\format{
  A data frame with 51 observations on the following 9 variables.
  \tabular{rl}{
    \code{Age} \tab {Age at diagnosis (in years)}\cr
    \code{Smear} \tab {Differential percentage of blasts}\cr
    \code{Infil} \tab {Percentage of absolute marrow leukemia infiltrate}\cr
    \code{Index} \tab {Percentage labeling index of the bone marrow leukemia cells}\cr
    \code{Blasts} \tab {Absolute number of blasts, in thousands}\cr
    \code{Temp} \tab {Highest temperature of the patient prior to treatment, in degrees Fahrenheit}\cr
    \code{Resp} \tab {\code{1}=responded to treatment or \code{0}=failed to respond}\cr
    \code{Time} \tab {Survival time from diagnosis (in months)}\cr
    \code{Status} \tab {\code{0}=dead or \code{1}=alive}\cr
  }
}
\details{
A study involved 51 untreated adult patients with acute myeloblastic
leukemia who were given a course of treatment, after which they were assessed as to their
response.
}
\source{
Data come from Statistical Analysis Using S-Plus (Brian S. Everitt; first edition 1994, Chapman & Hall).
}

\keyword{datasets}
