\name{LewyBody2Groups}
\alias{LewyBody2Groups}
\docType{data}
\title{Lewy Bodies and Dimentia}
\description{
Dementia study comparing two groups of patients
}
\format{
  A data frame with 39 observations on the following 3 variables.
  \describe{
    \item{\code{Type}}{\code{DLB}=Dementia with Lewy Bodies or \code{DLB/AD}=DLB and Alzheimer's Disease}
    \item{\code{APC}}{Annualized Percentage Change from baseline volume of the brain}
    \item{\code{MMSE}}{Change in functional performance on the Mini Mental State Examination}
  }
}
\details{
Brain MRIs were used to study the brains of patients with Dementia with Lewy Bodies, some of whom also were diagnosed with Alzheimer's Disease.
}
\source{
Z. Nedelksa et al. (2015), "Pattern of brain atrophy rates in autopsy-confirmed dementia with Lewy bodies,"" Neurobiology of Aging, 36: 452-461.
}
\keyword{datasets}
