\name{LewyDLBad}
\alias{LewyDLBad}
\docType{data}
\title{Lewy Bodies and Dimentia with Alzheimer's}
\description{
Dementia Study with Lewy Bodies
}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{group}}{\code{DLB/AD}=DLB and Alzheimer's Disease}
    \item{\code{APC}}{Annualized Percentage Change from baseline volume of the brain}
    \item{\code{MMSE}}{Change in functional performance on the Mini Mental State Examination}
  }
}
\details{
Brain MRIs were used to study the brains of patients with Dementia with Lewy Bodies. These are the cases that were also diagnosed with Alzheimer's Disease.  This is a subset of LewBody2Groups
}
\source{
Z. Nedelksa et al. (2015), "Pattern of brain atrophy rates in autopsy-confirmed dementia with Lewy bodies," Neurobiology of Aging, 36: 452-461.
}
\keyword{datasets}
