\name{LongJumpOlympics2016}
\alias{LongJumpOlympics2016}
\docType{data}
\title{Olympic Men's Long Jump Gold Medal Distance (1900 - 2016)}
\description{
Gold medal distance for Olympic men's long jump
}
\format{
  A data frame with 28 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Olympic Year (1900-2016)}
    \item{\code{Gold}}{Gold medal distance (in meters)}
  }
}
\details{
Gold medal winning distances for the men's long jump at the Olympics from 1900 to 2016.
}
\source{
Historical Olympic long jump results at http://trackandfield.about.com/od/longjump/qt/olymlongjumpmen.htm
}
\keyword{datasets}
