\name{LosingSleep}
\alias{LosingSleep}
\docType{data}
\title{Sleep Hours for Teenagers}
\description{
Hours of sleep for teenagers
}
\format{
  A data frame with 446 observations on the following 3 variables.
  \describe{
    \item{\code{Person}}{Cased ID number}
    \item{\code{Age}}{Age (in years)}
    \item{\code{Outcome}}{Average at least 7 hours of sleep? (\code{1=yes} or \code{0}=no)}
  }
}
\details{
Data from a random sample of 446 teens, aged 14 to 18, who answer the question, "On an average school night, how many hours of sleep do you get?" The outcome variable records whether or not each person averages at least 7 hours of sleep.
}
\source{
Need a reference.
}
\keyword{datasets}
