\name{Marathon}
\alias{Marathon}
\docType{data}
\title{Marathon}
\description{
Training records for a marathon runner
}
%\useage{data(Marathon)}
\format{
  A dataset with 1128 observations on the following 9 variables.
  \tabular{rl}{
    \code{Date} \tab {Training date}\cr
    \code{Miles} \tab {Miles for training run}\cr
    \code{Time} \tab {Training time (in minutes:seconds:hundredths)}\cr
    \code{Pace} \tab {Running pace (in minutes:seconds:hundredths per mile)}\cr
    \code{ShoeBrand} \tab {\code{Addidas}, \code{Asics}, \code{Brooks}, \code{Izumi}, \code{Mizuno}, or \code{New Balance}}\cr
    \code{TimeMin} \tab {Training time (in minutes)}\cr
    \code{PaceMin} \tab {Running pace (in minutes per mile)}\cr
    \code{Short} \tab {\code{1}= 5 miles or less or \code{0}=more than 5 miles}\cr
    \code{After2004} \tab {\code{1}= for runs after 2004 or \code{0}=for earlier runs}\cr
    }
}
\details{
Information from training records of a marathoner over a five-year period from 2002-2006. 
}
\source{
Data from training records of one of the Stat2 authors. 
}

\keyword{datasets}
