\name{MathEnrollment}
\alias{MathEnrollment}
\docType{data}
\title{Enrollments in Math Courses}
\description{
Semester enrollments in mathematics courses
}
%\useage{data(MathEnrollment)}
\format{
  A dataset with 11 observations on the following 3 variables.
  \tabular{rl}{
    \code{AYear} \tab {Academic year (for the fall)}\cr
    \code{Fall} \tab {Fall semester total enrollments}\cr
    \code{Spring} \tab {Spring semester total enrollments}\cr
  }
}
\details{
Total enrollments in mathematics courses at a
small liberal arts college were obtained for each semester from Fall 2001 to Spring 2012.
}
\source{
The data were obtained from http://Registrar.Kenyon.edu on June 1, 2012.
}

\keyword{datasets}
