\name{MathPlacement}
\alias{MathPlacement}
\docType{data}
\title{Math Placement Exam Results}
\description{
Results from a Math Placement exam at a liberal arts college
}
%\usage{data(MathPlacement)}
\format{
  A dataset with 2696 observations on the following 16 variables.
  \tabular{rl}{
    \code{Student} \tab {Identification number for each student}\cr
    \code{Gender} \tab {0=Female, 1=Male}\cr
    \code{PSATM} \tab {PSAT score in Math}\cr
    \code{SATM} \tab {SAT score in Math}\cr
    \code{ACTM} \tab {ACT Score in Math}\cr
    \code{Rank} \tab {Adjusted rank in HS class}\cr
    \code{Size} \tab {Number of students in HS class}\cr
    \code{GPAadj} \tab {Adjusted GPA}\cr
    \code{PlcmtScore} \tab {Score on math placement exam}\cr
    \code{Recommends} \tab {Recommended course: \code{R0} \code{R01} \code{R1} \code{R12} \code{R2} \code{R3} \code{R4} \code{R6} \code{R8}}\cr
    \code{Course} \tab {Actual course taken}\cr
    \code{Grade} \tab {Course grade}\cr
    \code{RecTaken} \tab {1=recommended course, 0=otherwise}\cr
    \code{TooHigh} \tab {1=took course above recommended, 0=otherwise}\cr
    \code{TooLow} \tab {1=took course below recommended, 0=otherwise}\cr
    \code{CourseSuccess} \tab {1=B or better grade, 0=grade below B}\cr
  }
}
\details{
Scores and course results for students taking a math placement exam at a college.
}
\source{
Personal correspondence
}

\keyword{datasets}
