\name{MedGPA}
\alias{MedGPA}
\docType{data}
\title{GPA and Medical School Admission}
\description{
Medical school admission status and information on GPA and standardized test scores
}
%\useage{data(MedGPA)}
\format{
  A dataset with 55 observations on the following 11 variables.
  \tabular{rl}{
    \code{Accept} \tab {Status: \code{A}=accepted to medical school or  \code{D}=denied admission}\cr
    \code{Acceptance} \tab {Indicator for Accept:  \code{1}=accepted or \code{0}=denied}\cr
    \code{Sex} \tab {\code{F}=female or  \code{M}=male}\cr
    \code{BCPM} \tab {Bio/Chem/Physics/Math grade point average}\cr
    \code{GPA} \tab {College grade point average}\cr
    \code{VR} \tab {Verbal reasoning (subscore)}\cr
    \code{PS} \tab {Physical sciences (subscore)}\cr
    \code{WS} \tab {Writing sample (subcore)}\cr
    \code{BS} \tab {Biological sciences (subscore)}\cr
    \code{MCAT} \tab {Score on the MCAT exam (sum of CR+PS+WS+BS)}\cr
    \code{Apps} \tab {Number of medical schools applied to}\cr
  }
}
\details{
This dataset has information gathered on 55 medical school applicants
from a liberal arts college in the Midwest.
}
\source{
Data collected at a midwestern liberal arts college.
}

\keyword{datasets}
