\name{Meniscus}
\alias{Meniscus}
\docType{data}
\title{Meniscus Repair Methods}
\description{
Comparing meniscus repair methods on cadaver knees
}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{\code{Method}}{Meniscus repair method (\code{1} = Vertical Suture, \code{2} = Meniscus Arrow, \code{3} = FasT-Fix)}
    \item{\code{FailureLoad}}{Load at failure (in Newtons)}
    \item{\code{Displacement}}{Displacement (in mm)}
    \item{\code{Stiffness}}{Stiffness (Newtons/mm)}
  }
}
\details{
Eighteen, lightly embalmed, cadaveric knee specimens were used in a study to compare three different methods of meniscus repair.  The specimens were randomly assigned to one of the three treatments: vertical suture, meniscus arrow, FasT-Fix.  They were evaluated on three different response variables: load at failure, stiffness, and displacement.
}
\source{
P. Borden, J. Nyland, D.N.M. Caborn, D. Pienkowski (2003), "Biomechanical Comparison of the FasT-Fix Meniscal Repair Suture System with Vertical Mattress Sutures and Meniscus Arrows," The American Journal of Sports Medicine, Vol. 31, #3, pp. 374-378

Dataset downloaded from http://www.stat.ufl.edu/~winner/data/meniscus.txt
}
\keyword{datasets}
