\name{MentalHealth}
\alias{MentalHealth}
\docType{data}
\title{Mental Health Admissions}
\description{
Admissions to a mental health emergency room and full moons
}
%\usage{data(MentalHealth)}
\format{
  A dataset with 36 observations on the following 3 variables.
  \tabular{rl}{
    \code{Month} \tab {Month of the year}\cr
    \code{Moon} \tab {Relationship to full moon: \code{After}, \code{Before}, or \code{During}}\cr
    \code{Admission} \tab {Number of emergency room admissions}\cr
  }
}
\details{
Some researchers in the early 1970s set out to study whether there is a
"full-moon" effect on emergency room admissions at a mental health hospital. They separated the data over 12 months into
rates before the full moon (mean number of patients seen 4-13 days before the full moon), during
the full moon (the number of patients seen on the full moon day), and after the full moon (mean
number of patients seen 4-13 days after the full moon).
}
\source{
Introduction to Mathematical Statistics and its Applications by Richard J. Larsen and Morris L. Marx.
Prentice Hall:Englewood Cliffs, NJ, 1986.
}
\references{
The original discussion of the study is in Blackman, S., and Catalina, D. (1973). "The moon and the emergency
room." Perceptual and Motor Skills 37, 624-626.
}

\keyword{datasets}
