\name{MetabolicRate}
\alias{MetabolicRate}
\docType{data}
\title{Metabolic Rate of Caterpillars}
\description{
Body size and metabolic rate of Manduca Sexta caterpillars
}
%\useage{data(MetabolicRate)}
\format{
  A dataset with 305 observations on the following 7 variables.
  \tabular{rl}{
    \code{Computer} \tab {ID number of the computer used to measure metabolic rate}\cr
    \code{BodySize} \tab {Size of the caterpillar (in grams)}\cr
    \code{LogBodySize} \tab {Log (base 10) of BodySize}\cr
    \code{Instar} \tab {Number from 1 (smallest) to 5 (largest) indicating stage of the caterpillar's life}\cr
    \code{CO2ppm} \tab {Carbon dioxide concentration (in ppm)}\cr
    \code{Mrate} \tab {Metabolic rate}\cr
    \code{LogMrate} \tab {Log (base 10) of metabolic rate}\cr
  }
}
\details{
Marisa Stearns collected and analyzed body size and
metabolic rates for Manduca Sexta caterpillars.
}
\source{
We thank Professor Itagaki and his research students for sharing these data.
}

\keyword{datasets}
