\name{MetroCommutes}
\alias{MetroCommutes}
\docType{data}
\title{Commute Times}
\description{
Commute times for four cities
}
\format{
  A data frame with 2000 observations on the following 3 variables.
  \describe{
    \item{\code{City}}{\code{Boston}, \code{Houston}, \code{Minneapolis}, or \code{Washington}}
    \item{\code{Distance}}{Distance of commute (in miles)}
    \item{\code{Time}}{Time of commute (in minutes)}
  }
}
\details{
The data are distances (miles) and times (minutes) of daily commute (one-way) for random samples of 500 commuters in each of four cities (Boston, Houston, Minneapolis, Washington) in 2007.  The random samples were taken from the Metropolitan Public Use File of the 2007 American Housing Survey
}
\source{
2007 American Housing Survey
https://www.census.gov/programs-surveys/ahs/data/2007/ahs-2007-public-use-file--puf-.html
}
\keyword{datasets}
