\name{MetroHealth83}
\alias{MetroHealth83}
\docType{data}
\title{Health Services in Metropolitan Areas}
\description{
Health services data for 83 metropolitan areas
}
%\useage{data(MetroHealth83)}
\format{
  A dataset with 83 observations on the following 16 variables.
  \tabular{rl}{
    \code{City} \tab {Name of the metropolitan area}\cr
    \code{NumMDs} \tab {Number of physicians}\cr
    \code{RateMDs} \tab {Number of physicians per 100,000 people}\cr
    \code{NumHospitals} \tab {Number of community hospitals}\cr
    \code{NumBeds} \tab {Number of hospital beds}\cr
    \code{RateBeds} \tab {Number of hospital beds per 100,000 people}\cr
    \code{NumMedicare} \tab {Number of Medicare recipients in 2003}\cr
    \code{PctChangeMedicare} \tab {Percent change in Medicare recipients (2000 to 2003)}\cr
    \code{MedicareRate} \tab {Number of Medicare recipients per 100,000 people}\cr
    \code{SSBNum} \tab {Number of Social Security recipients in 2004}\cr
    \code{SSBRate} \tab {Number of Social Security recipients per 100,000 people}\cr
    \code{SSBChange} \tab {Percent change in Social Security recipients (2000 to 2004)}\cr
    \code{NumRetired} \tab {Number of retired workers}\cr
    \code{SSINum} \tab {Number of Supplemental Security Income recipients in 2004}\cr
    \code{SSIRate} \tab {Number of Supplemental Security Income recipients per 100,000 people}\cr
    \code{SqrtMDs} \tab {Square root of number of physicians}\cr
  }
}
\details{
The U.S. Census Bureau regularly collects information for many metropolitan areas in the United
States, including data on number of physicians and number (and size) of hospitals.
This dataset has such information for 83 different metropolitan areas.

This dataset is in the first edition, but replaced by CountyHealth in the second edition.
}
\source{
U.S. Census Bureau: 2006 State and Metropolitan Area Data Book (Table B-6) \cr
http://www.census.gov/prod/2006pubs/smadb/smadb-06.pdf
}

\keyword{datasets}
