\name{MothEggs}
\alias{MothEggs}
\docType{data}
\title{Moth Eggs}
\description{
Body size and eggs produced for a species of moths
}
%\useage{data(MothEggs)}
\format{
  A dataset with 39 observations on the following 2 variables.
  \tabular{rl}{
    \code{BodyMass} \tab {Log of body size measured in grams}\cr
    \code{Eggs} \tab {Number of eggs present}\cr
  }
}
\details{
Researchers were interested in an association between body size and the number of eggs produced
by a species of moths.
}
\source{
We thank Professor Itagaki and his students for sharing this data from experiments on Manduca Sexta.
}

\keyword{datasets}
