\name{NFL2007Standings}
\alias{NFL2007Standings}
\docType{data}
\title{NFL Standings for 2007 Regular Season}
\description{
Standings for National Football League teams in 2007
}
%\useage{data(NFL2007Standings)}
\format{
  A dataset with 32 observations on the following 10 variables.
  \tabular{rl}{
    \code{Team} \tab {Team name}\cr
    \code{Conference} \tab {Conference: \code{AFC} or \code{NFC}}\cr
    \code{Division} \tab {Division within conference: \code{ACE}, \code{ACN}, \code{ACS}, \code{ACW}, \code{NCE}, \code{NCN}, \code{NCS}, \code{NCW}}\cr
    \code{Wins} \tab {Number of wins (out of 16 games)}\cr
    \code{Losses} \tab {Number of losses}\cr
    \code{WinPct} \tab {Proportion of games won (Wins/16)}\cr
    \code{PointsFor} \tab {Total points scored by the team}\cr
    \code{PointsAgainst} \tab {Total points scored against the team}\cr
    \code{NetPts} \tab {PointsFor minus PointsAgainst}\cr
    \code{TDs} \tab {Number of touchdowns scored by the team}\cr
  }
}
\details{
Data for all 32 National Football League (NFL) teams for the 2007 regular season.

Updated to NFLStandings2016 in the second edition.
}
\source{
Data downloaded from www.nfl.com
}

\keyword{datasets}
