\name{NFLStandings2016}
\alias{NFLStandings2016}
\docType{data}
\title{NFL Standings for 2016 Regular Season}
\description{
Standings and team statistics for National Football League (NFL) teams in the 2016 season
}
\format{
  A data frame with 32 observations on the following 11 variables.
  \describe{
    \item{\code{Team}}{Team name}
    \item{\code{Wins}}{Wins in the 2016 regular season (out of 16 games)}
    \item{\code{Losses}}{Losses in the 2016 regular season}
    \item{\code{Ties}}{Ties in the 2016 regular season (ties are very rare in the NFL)}
    \item{\code{WinPct}}{Winning percentage = (Wins+0.5*Ties)/16 games}
    \item{\code{PointsFor}}{Points scored}
    \item{\code{PointsAgainst}}{Points allowed}
    \item{\code{NetPts}}{Points scored minus Points allowed}
    \item{\code{YardsFor}}{Offensive yards gained by the team}
    \item{\code{YardsAgainst}}{Offensive yards against the team}
    \item{\code{TDs}}{Touchdowns scored}
  }
}
\details{
Standings for the 2016 regular season of the National Football League (NFL) along with points and scored and allowed for each team in its16 games.
}
\source{
Data downloaded from: \cr
http://www.pro-football-reference.com/years/2016/
}
\keyword{datasets}
