\name{Nursing}
\alias{Nursing}
\docType{data}
\title{Nursing Homes}
\description{
Characteristics of nursing homes in New Mexico.
}
%\useage{data(Nursing)}
\format{
  A dataset with 52 observations on the following 7 variables.
  \tabular{rl}{
    \code{Beds} \tab {Number of beds in the nursing home}\cr
    \code{InPatientDays} \tab {Annual medical in-patient days (in hundreds)}\cr
    \code{AllPatientDays} \tab {Annual total patient days (in hundreds)}\cr
    \code{PatientRevenue} \tab {Annual patient care revenue (in hundreds of dollars)}\cr
    \code{NurseSalaries} \tab {Annual nursing salaries (in hundreds of dollars)}\cr
    \code{FacilitiesExpend} \tab {Annual facilities expenditure (in hundreds of dollars)}\cr
    \code{Rural} \tab {1=\code{rural} or 0=\code{non-rural}}\cr
  }
}
\details{
The data were collected by the Department of Health and Social Services of the
State of New Mexico and cover 52 of the 60 licensed nursing facilities in New Mexico
in 1988.
}
\source{
Downloaded from DASL at http://lib.stat.cmu.edu/DASL/Datafiles/Nursingdat.html
}
\references{
Howard L. Smith, Niell F. Piland, and Nancy Fisher,
"A Comparison of Financial Performance, Organizational Characteristics,
and Management Strategy Among Rural and Urban Nursing Facilities,"
Journal of Rural Health, Winter 1992, pp 27-40.
}

\keyword{datasets}
