\name{Olives}
\alias{Olives}
\docType{data}
\title{Fenthion in Olive Oil}
\description{
Measurements of the pesticide fenthion in olive oil over time
}
%\useage{data(Olives)}
\format{
  A dataset with 18 observations on the following 7 variables.
  \tabular{rl}{
    \code{SampleNumber} \tab {Code (1-6) for sample of olive oil}\cr
    \code{Group} \tab {Code for group: \code{1} or \code{2}}\cr
    \code{Day} \tab {Time (in days) when sample was measured: \code{0}, \code{281}, or \code{365}}\cr
    \code{Fenthion} \tab {Amount of fenthion (pesticide)}\cr
    \code{FenthionSulphoxide} \tab {Amount of fenthion sulfide}\cr
    \code{FenthionSulphone} \tab {Amount of fenthion sulphone}\cr
    \code{Time} \tab {Code (0, 3, or 4) for the number of days}\cr
  }
}
\details{
Fenthion is a pesticide used against the olive fruit fly in olive groves. It is toxic
to humans so it is important that there be no residue left on the fruit or in olive oil that will be
consumed. One theory was that if there is residue of the pesticide left in the olive oil, it would
dissipate over time. Chemists set out to test that theory by taking a random sample of small
amounts of olive oil with fenthion residue and measuring the amount of fenthion in the oil at three
different times over the year - day 0, day 281 and day 365.
}
\source{
Data provided by Rosemary Roberts and discussed in "Persistence of fenthion residues in olive oil" by Chaido
Lentza-Rizos, Elizabeth J. Avramides, and Rosemary A. Roberts in Pest Management Science, Vol. 40, Issue 1, Jan.
1994, pp. 63-69.
}

\keyword{datasets}
