\name{Overdrawn}
\alias{Overdrawn}
\docType{data}
\title{Overdrawn Checking Account?}
\description{
Survey of college students to took at factors related to having overdrawn a checking account.
}
\format{
  A dataset with 450 observations on the following 4 variables.
  \tabular{rl}{
    \code{Age} \tab {Age of the student (in years)}\cr
    \code{Sex} \tab {\code{0}=male or \code{1}=female}\cr
    \code{DaysDrink} \tab {Number of days drinking alcohol (in past 30 days)}\cr
    \code{Overdrawn} \tab {Has student overdrawn a checking account? \code{0}=no or \code{1}=yes}\cr
  }
}
\details{
Researchers conducted a survey of 450 undergraduates in large
introductory courses at either Mississippi State University or the University of Mississippi. There
were close to 150 questions on the survey, but only four of these variables are included in this
dataset. (You can consult the paper to learn how the variables beyond these 4 affect
the analysis.)  The primary interest for the researchers was factors relating to whether or not a student
has ever overdrawn a checking account.

Renamed as CreditRisk in second edition.
}
\source{
Worthy S.L., Jonkman J.N., Blinn-Pike L. (2010), "Sensation-Seeking, Risk-Taking, and Problematic Financial Behaviors of College Students,"
Journal of Family and Economic Issues, 31: 161-170
}

\keyword{datasets}
