\name{PeaceBridge2003}
\alias{PeaceBridge2003}
\docType{data}
\title{Monthly Peace Bridge Traffic ( 2003-2015)}
\description{
Monthly traffic (in 1,000's) across the Peace Bridge between Canada and the U.S.
}
\format{
  A data frame with 156 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (2003 to 2015)}
    \item{\code{Month}}{Month (1 to 12)}
    \item{\code{Traffic}}{Vehicles (in 1,000's)}
    \item{\code{t}}{Time frame (1 to 156)}
  }
}
\details{
Monthly traffic (in thousands of vehicles) across the Peace Bridge between the U.S. and Canada near Niagara Falls between January 2003 and December 2015.  Note PeaceBridge2012 has only the last four years of this series.
}
\source{
http://www.peacebridge.com/index.php/historical-traffic-statistics/yearly-volumes
}
\keyword{datasets}
