\name{PeaceBridge2012}
\alias{PeaceBridge2012}
\docType{data}
\title{Monthly Peace Bridge Traffic ( 2012-2015)}
\description{
Monthly traffic (in 1,000's) across the Peace Bridge between Canada and the U.S.
}
\format{
  A data frame with 48 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (2012 to 2015)}
    \item{\code{Month}}{Month (1 to 12)}
    \item{\code{Traffic}}{Vehicles (in 1,000's)}
    \item{\code{t}}{Time frame (1 to 48)}
  }
}
\details{
Monthly traffic (in thousands of vehicles) across the Peace Bridge between the U.S. and Canada near Niagara Falls between January 2012 and December 2015.  Note PeaceBridge2003 has similar data starting in 2003.
}
\source{
http://www.peacebridge.com/index.php/historical-traffic-statistics/yearly-volumes
}
\keyword{datasets}
