\name{Pedometer}
\alias{Pedometer}
\docType{data}
\title{Pedometer Walking Data}
\description{
Daily walking amounts recorded on a personal pedometer from September-December 2011
}
\format{
  A dataset with 68 observations on the following 8 variables.
  \tabular{rl}{
    \code{Steps} \tab {Total number of steps for the day}\cr
    \code{Moderate} \tab {Number of steps at a moderate walking speed}\cr
    \code{Min} \tab {Number of minutes walking at a moderate speed}\cr
    \code{kcal} \tab {Number of calories burned walking at a moderate speed}\cr
    \code{Mile} \tab {Total number of miles walked}\cr
    \code{Rain} \tab {Type of weather (\code{rain} or \code{shine})}\cr
    \code{Day} \tab {Day of the week (\code{U}=Sunday, \code{M}=Monday,
\code{T}=Tuesday, \code{W}=Wednesday, \code{R}=Thursday, \code{F}=Friday, \code{S}=Saturday}\cr
    \code{DayType} \tab {Coded as \code{Weekday} or \code{Weekend}}\cr
  }
}
\details{
A statistics professor regularly keeps a pedometer in his pocket. It records
not only the number of steps taken each day, but also the number of steps taken at a moderate
pace, the number of minutes walked at a moderate pace, and the number of miles total that he
walked. He also added to the data set the day of the week, whether it was rainy, sunny, or cold (on
sunny days he often biked, but on rainy or cold days he did not), and whether it was a weekday
or weekend.
}
\source{
One of the Stat2 authors
}

\keyword{datasets}
