\name{Perch}
\alias{Perch}
\docType{data}
\title{Perch Sizes}
\description{
Size of perch caught in a Finnish lake
}
%\useage{data(Perch)}
\format{
  A dataset with 56 observations on the following 4 variables.
  \tabular{rl}{
    \code{Obs} \tab {Observation number}\cr
    \code{Weight} \tab {Weight (in grams)}\cr
    \code{Length} \tab {Length (in centimeters)}\cr
    \code{Width} \tab {Width (in centimeters}\cr
  }
}
\details{
This dataset comes from a sample of fish (perch) caught at Lake Laengelmavesi in Finland.
}
\source{
JSE Data Archive, http://www.amstat.org/publications/jse/jse_data_archive.htm, submitted by Juha Puranen.
}

\keyword{datasets}
