\name{PigFeed}
\alias{PigFeed}
\docType{data}
\title{Additives in Pig Feed}
\description{
Effects of additives to pig feed on weight gain
}
\format{
  A dataset with 12 observations on the following 3 variables.
  \tabular{rl}{
    \code{WgtGain} \tab {Daily weight gain (hundredths of a pound over 1.00)}\cr
    \code{Antibiotic} \tab {Antibiotic in the feed? \code{No} or \code{Yes}}\cr
    \code{B12} \tab {Vitamin B12 in the feed? \code{No} or \code{Yes}}\cr
  }
}
\details{
A scientist in Iowa was interested in additives to standard pig chow that might increase the rate at
which the pigs gained weight.  Two factors of interest were vitamin B12 and antibiotics.
To perform the experiment, the scientist randomly assigned 12 pigs, three
to each of the diet combinations (Antibiotic only, B12 only, both, and neither).
}
\source{
Data are found in Statistical Methods by George W. Snedecor and William G. Cochran (1967). Ames, IA:
The Iowa State University Press.
}
\references{
Original source is Iowa Agricultural Experiment Station (1952). Animal Husbandry Swine Nutrition Experiment
No. 577.
}

\keyword{datasets}
