\name{Pines}
\alias{Pines}
\docType{data}
\title{Measurements of Pine Tree Seedlings}
\description{
Data from pine seedlings planted in 1990
}
\format{
  A dataset with 1000 observations on the following 15 variables.
  \tabular{rl}{
    \code{Row} \tab {Row number in pine plantation}\cr
    \code{Col} \tab {Column number in pine plantation}\cr
    \code{Hgt90} \tab {Tree height at time of planting (cm)}\cr
    \code{Hgt96} \tab {Tree height in September 1996 (cm)}\cr
    \code{Diam96} \tab {Tree trunk diameter in September 1996 (cm)}\cr
    \code{Grow96} \tab {Leader growth during 1996 (cm)}\cr
    \code{Hgt97} \tab {Tree height in September 1997 (cm)}\cr
    \code{Diam97} \tab {Tree trunk diameter in September 1997 (cm)}\cr
    \code{Spread97} \tab {Widest lateral spread in September 1997 (cm)}\cr
    \code{Needles97} \tab {Needle length in September 1997 (mm)}\cr
    \code{Deer95} \tab {Type of deer damage in September 1995: 0 = none, 1 = browsed}\cr
    \code{Deer97} \tab {Type of deer damage in September 1997: 0 = none, 1 = browsed}\cr
    \code{Cover95} \tab {Thorny cover in September 1995: 0 = none; 1 = some; 2 = moderate; 3 = lots}\cr
    \code{Fert} \tab {Indicator for fertilizer: 0 = no, 1 = yes}\cr
    \code{Spacing} \tab {Distance (in feet) between trees (10 or 15)}\cr
  }
}
\details{
This dataset contains data from an experiment conducted by the Department
of Biology at Kenyon College at a site near the campus in Gambier, Ohio. In April 1990, student
and faculty volunteers planted 1000 white pine (Pinus strobes) seedlings at the Brown Family Environmental
Center. These seedlings were planted in two grids, distinguished by 10- and 15-foot
spacings between the seedlings. Several variables were measured and recorded
for each seedling over time (in 1990, 1996, and 1997).
}
\source{
Thanks to the Kenyon College Department of Biology for sharing these data.
}

\keyword{datasets}
