\name{PKU}
\alias{PKU}
\docType{data}
\title{Dopamine levels with PKU in diets}
\description{
Dopamine levels with different amounts of phenylalanine in diets
}
\format{
  A data frame with 20 observations on the following 4 variables.
  \describe{
    \item{\code{Subject}}{Initials to identify each subject}
    \item{\code{Diet}}{Level of phenylalanine in diet (\code{Low} or \code{Normal})}
    \item{\code{DietControl}}{Ability to follow prescribed diet  (\code{Good} or \code{Poor})}
    \item{\code{Y}}{Concentration of dopamine (micrograms per milligram of creatinine)}
  }
}
\details{
Phenylketonuria (PKU) is an enzyme deficiency that keeps a person from being able to synthesize enough dopamine. The amino acid phenylalanine inhibits the enzyme needed to synthesize dopamine, and so to some extent, a diet low in phenylalanine can moderate the symptoms of PKU. In short, less phenylalanine in the diet should lead to more dopamine in the brain. The dopamine level for each patient was measured after a normal diet and after a week on a low phenylalanine diet.
}
\source{
Krause, Halminski, McDonald, Dembure, Salvo, Freides, and Elsas (1985) "Biochemical and Neuropsychological Effects of Elevated Plasma Phenylalanine in Patients with Treated Phenylketonuria," J. of Clinical Investigation, Volume 75, January 1985, 40-48

Several of the values were altered slightly in ways that would not change the analysis except to simplify the arithmetic.

}
\keyword{datasets}
