\name{Political}
\alias{Political}
\docType{data}
\title{Political Behavior of College Students}
\description{
Survey of political activity for Grinnell College students
}
\format{
  A dataset with 59 observations on the following 9 variables.
  \tabular{rl}{
    \code{Year} \tab {Class year (1 to 4)}\cr
    \code{Sex} \tab {\code{0}=male or \code{1}=female}\cr
    \code{Vote} \tab {Voting status: \code{0}=not eligible, \code{1}=eligible/not registered, \code{2}=registered/didn't vote, \code{4}=voted}\cr
    \code{Paper} \tab {Read news (per week): \code{0}=never, \code{1}=less than once, \code{2}=once, \code{3}=2 or 3 times, \code{4}=daily}\cr
    \code{Edit} \tab {Read editorial page? \code{0}=no or \code{1}=yes}\cr
    \code{TV} \tab {Watch TV news: \code{0}=never, \code{1}=less than once, \code{2}=once, \code{3}=2 or 3 times, \code{4}=daily}\cr
    \code{Ethics} \tab {Politics should be ruled by: \code{1}=ethical considerations to \code{5}=practical power}\cr
    \code{Inform} \tab {How informed are you about politics?  \code{1}=uninformed to \code{5}=very well informed}\cr
    \code{Participate} \tab {Missing if Vote=0, \code{0} if Vote=1 or 2, \code{1} if Vote=3}\cr
  }
}
\details{
Students Jennifer Wolfson and Meredith Goulet conducted
a survey in the spring of 1992 of Grinnell College students to ascertain patterns of political behavior.
They took a simple random sample of 60 students who were U.S. citizens and conducted
phone interviews. Using several "call backs" they obtained 59 responses.
}
\source{
Student survey at Grinnell College
}

\keyword{datasets}
