\name{Pollster08}
\alias{Pollster08}
\docType{data}
\title{Pollster08}
\description{
Polls for 2008 U.S. presidential election
}
%\useage{data(Pollster08)}
\format{
  A dataset with 102 observations on the following 11 variables.
  \tabular{rl}{
    \code{PollTaker} \tab {Polling organization}\cr
    \code{PollDates} \tab {Dates the poll data were collected}\cr
    \code{MidDate} \tab {Midpoint of the polling period}\cr
    \code{Days} \tab {Number of days after August 28th (end of Democratic convention)}\cr
    \code{n} \tab {Sample size for the poll}\cr
    \code{Pop} \tab {\code{A}=all,  \code{LV}=likely voters, \code{RV}=registered voters}\cr
    \code{McCain} \tab {Percent supporting John McCain}\cr
    \code{Obama} \tab {Percent supporting Barak Obama}\cr
    \code{Margin} \tab {Obama percent minus McCain percent}\cr
    \code{Charlie} \tab {Indicator for polls after Charlie Gibson interview with VP candidate Sarah Palin (9/11)}\cr
    \code{Meltdown} \tab {Indicator for polls after Lehman Brothers bankruptcy (9/15)}\cr
  }
}
\details{
The file Pollster08 contains data from 102 polls that
were taken during the 2008 U.S. Presidential campaign. These data include all presidential polls
reported on the internet site pollster.com that were taken between August 29th, when John Mc-
Cain announced that Sarah Palin would be his running mate as the Republican nominee for vice
president, and the end of September.
}
\source{
Downloaded from pollster.com
}

\keyword{datasets}
