\name{Popcorn}
\alias{Popcorn}
\docType{data}
\title{Popcorn Popping Success}
\description{
Unpopped kernels in bags of microwave popcorn
}
\format{
  A dataset with 12 observations on the following 3 variables.
  \tabular{rl}{
    \code{Unpopped} \tab {Number of unpopped kernels (adjusted for size difference)}\cr
    \code{Brand} \tab {\code{Orville} or \code{Seaway}}\cr
    \code{Trial} \tab {Trial number}\cr
  }
}
\details{
Two students, Lara and Lisa, conducted an experiment to compare
Orville Redenbacher's Light Butter Flavor vs. Seaway microwave popcorn. They made 12 batches
of popcorn, 6 of each type, cooking each batch for four minutes. They noted that the microwave
oven seemed to get warmer as they went along so they kept track of six trials and randomly
chose which brand would go first for each trial. For a response variable they counted the number
of unpopped kernels and then adjusted the count for Seaway for having more ounces per bag of
popcorn (3.5 vs 3.0).
}
\source{
Student project
}

\keyword{datasets}
