\name{PorscheJaguar}
\alias{PorscheJaguar}
\docType{data}
\title{Porsche and Jaguar Prices}
\description{
Compare prices for Porsche and Jaguar cars offered for sale at an internet site
}
\format{
  A dataset with 60 observations on the following 5 variables.
  \tabular{rl}{
    \code{Car} \tab {Car model: \code{Jaguar} or  \code{Porsche}}\cr
    \code{Price} \tab {Price (in $1,000's)}\cr
    \code{Age} \tab {Age of the car (in years)}\cr
    \code{Mileage} \tab {Previous miles driven (in 1,000's)}\cr
    \code{Porsche} \tab {Indicator for Porsche (\code{1}) or Jaguar (\code{0})}\cr
  }
}
\details{
Two students collected samples of Porsche and Jaguar cars that were offered for sale at an internet site.
In addition to asking price, they recorded the model year (converting to age) and mileage of each advertised car.
}
\source{
Student project data collected from autotrader.com in Spring 2007.
}

\keyword{datasets}
