\name{PorschePrice}
\alias{PorschePrice}
\docType{data}
\title{Porsche Prices}
\description{
Prices for Porsche cars offered for sale at an internet site.
}
\format{
  A dataset with 30 observations on the following 3 variables.
  \tabular{rl}{
    \code{Price} \tab {Asking price for the car (in $1,000's)}\cr
    \code{Age} \tab {Age of the car (in years)}\cr
    \code{Mileage} \tab {Previous miles driven (in 1,000's)}\cr
  }
}
\details{
A student was interested in prices for used Porsche sports cars being sold on the internet.
He selected a random sample of 30 Porsches from the ones being advertised at autotrader.com.
For each car he recorded the asking price, mileage, and model year (which he converted to age).

This dataset was replaced by AccordPrice for second edition.
}
\source{
Data collected for a student project from autotrader.com in February 2007.
}

\keyword{datasets}
