\name{Pulse}
\alias{Pulse}
\docType{data}
\title{Pulse Rates and Exercise}
\description{
Pulse rates before and after exercise for a sample of statistics students
}
%\useage{data(Pulse)}
\format{
  A dataset with 232 observations on the following 7 variables.
  \tabular{rl}{
    \code{Active} \tab {Pulse rate (beats per minute) after exercise}\cr
    \code{Rest} \tab {Resting pulse rate (beats per minute)}\cr
    \code{Smoke} \tab {\code{1}=smoker or \code{0}=nonsmoker}\cr
    \code{Sex} \tab {\code{1}=female or \code{0}=male}\cr
    \code{Exercise} \tab {Typical hours of exercise (per week)}\cr
    \code{Hgt} \tab {Height (in inches)}\cr
    \code{Wgt} \tab {Weight (in pounds)}\cr
  }
}
\details{
Students in a Stat2 class recorded resting pulse rates (in class),
did three "laps" walking up/down a nearby set of stairs, and then
measured their pulse rate after the exercise.  They provided additional information about
height, weight, exercise, and smoking habits via a survey.
}
\source{
Data compiled over several semesters from students taking a Stat2 course.
}

\keyword{datasets}
