\name{Putts1}
\alias{Putts1}
\docType{data}
\title{Putting Success by Length (Long Form)}
\description{
Putting results for a golfing statistician
}
\format{
  A dataset with 587 observations on the following 2 variables.
  \tabular{rl}{
    \code{Length} \tab {Length of the putt (in feet)}\cr
    \code{Made} \tab {\code{1}=made the putt or \code{0}=missed the putt}\cr
  }
}
\details{
A statistician golfer kept careful records of every putt he attempted when playing golf,
recording the length of the putt and whether or not he was successful in making the putt.
This dataset has one case for each of the 587 attempted putts.
 A different form of the same data (Putts2) accumulates counts of makes and misses for each putt length.
}
\source{
Personal observations by one of the Stat2 authors
}

\keyword{datasets}
