\name{Putts2}
\alias{Putts2}
\docType{data}
\title{Putting Success by Length (Short Form)}
\description{
Putting results for a golfing statistician (by length of the putts)
}
%\useage{data(Putts2)}
\format{
  A dataset with 5 observations on the following 4 variables.
  \tabular{rl}{
    \code{Length} \tab {Length of the attempted putt (in feet)}\cr
    \code{Made} \tab {Number of putts made at this length}\cr
    \code{Missed} \tab {Number of putts missed at this length}\cr
    \code{Trials} \tab {Total number of putts attempted at this length}\cr
  }
}
\details{
A statistician golfer kept careful records of every putt he attempted when playing golf,
recording the length of the putt and whether or not he was successful in making the putt.
For each different length, this dataset records the number of putts made, missed, and the total number
of attempts from that length. A similar dataset, Putts1, has one case for each of the 587 attempted putts,
showing the length and outcome.
}

\source{
Personal observations by one of the Stat2 authors
}

\keyword{datasets}
