\name{Putts3}
\alias{Putts3}
\docType{data}
\title{Hypothetical Putting Data (Short Form)}
\description{
Hypothetical putting results for a golfing statistician
}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{Length}}{Length of the attempted putt (in feet)}
    \item{\code{Made}}{Number of putts made at this length}
    \item{\code{Missed}}{Number of putts missed at this length}
    \item{\code{Trials}}{Total number of putts attempted at this length}
  }
}
\details{
This is a hypothetical revision of the table of putting success in Putts2 that helps demonstrate overdispersion.
}
\source{
Modified from personal observations by one of the Stat2 authors.
}
\keyword{datasets}
