\name{Rectangles}
\alias{Rectangles}
\docType{data}
\title{Measurements of Rectangles}
\description{
Measurements for a hypothetical set of nine rectangles.
}
\format{
  A data frame with 9 observations on the following 5 variables.
  \describe{
    \item{\code{Case}}{ID number for each rectangle}
    \item{\code{Width}}{Width (1, 4, or 10)}
    \item{\code{Length}}{Length (1, 4, or 10)}
    \item{\code{Area}}{Area}
    \item{\code{logArea}}{Log (base 10) of area}
  }
}
\details{
Areas for rectangles of width 1, 4, or 10 and length of 1, 4, or 10.
}
\source{
Areas computed for a hypothetical set of rectangles.
}
\keyword{datasets}
