\name{ReligionGDP}
\alias{ReligionGDP}
\docType{data}
\title{Religion and GDP for Countries}
\description{
Data on religiosity of countries from the Pew Global Attitudes Project
}
%\useage{data(ReligionGDP)}
\format{
  A dataset with 44 observations on the following 9 variables.
  \tabular{rl}{
    \code{Country} \tab {Name of country}\cr
    \code{Religiosity} \tab {A measure of degree of religiosity for residents of the country}\cr
    \code{GDP} \tab {Per capita Gross Domestic Product in the country}\cr
    \code{Africa} \tab {Indicator for countries in Africa }\cr
    \code{EastEurope} \tab {Indicator for countries in Eastern Europe }\cr
    \code{MiddleEast} \tab {Indicator for countries in the Middle East }\cr
    \code{Asia} \tab {Indicator for countries in Asia }\cr
    \code{WestEurope} \tab {Indicator for countries in Western Europe }\cr
    \code{Americas} \tab {Indicator for countries in North/South America}\cr
  }
}
\details{
The Pew Research Center's Global Attitudes Project surveyed
people around the world and asked (among many other questions) whether they agreed that "belief
in God is necessary for morality," whether religion is very important in their lives, and whether
they pray at least once per day. The variable Religiosity is the sum of the percentage of positive
responses on these three items, measured in each of 44 countries. The dataset also includes the
per capita GDP for each country and indicator variables
that record the part of the world the country is in.
}
\source{
Data from the 2007 Spring Survey conducted through the
Pew Global Attitudes Project at http://www.pewglobal.org.
}

\keyword{datasets}
