\name{RepeatedPulse}
\alias{RepeatedPulse}
\docType{data}
\title{Pulse Rates at Various Times of Day}
\description{
A student measured her pulse several times a day over 26 days.
}
\format{
  A data frame with 104 observations on the following 3 variables.
  \describe{
    \item{\code{Pulse}}{Pulse rate (beats per minute)}
    \item{\code{Time}}{Time of day (\code{evening}, \code{morning}, \code{noon}, \code{one})}
    \item{\code{Day}}{ \code{Day1} to  \code{Day26} }
  }
}
\details{
 A student measured her pulse in the morning, at noon, at 1:00, and in the evening for each of 26 days.
}
\source{
Data supplied by a student at Oberlin College.
}
\keyword{datasets}
