\name{Retirement}
\alias{Retirement}
\docType{data}
\title{Yearly Contributions to a Supplemental Retirement Account}
\description{
Contributions to a supplemental retirement account (1997-2012)
}
%\useage{data(Retirement)}
\format{
  A dataset with 16 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {1997-2012}\cr
    \code{SRA} \tab {Annual contribution to the Supplemental Retirement Account}\cr
  }
}
\details{
A faculty member opened a supplemental retirement account (SRA) in 1997 to
invest money for retirement. This dataset shows the annual contributions to that account.
Annual contributions were adjusted downward during sabbatical
years in order to maintain a steady family income.
}
\source{
Individual records kept by the faculty member.
}

\keyword{datasets}
