\name{RiverElements}
\alias{RiverElements}
\docType{data}
\title{Elements in River Water Samples}
\description{
Concentrations of elements in river water samples from upstate NY
}
%\useage{data(RiverElements)}
\format{
  A dataset with 12 observations on the following 27 variables.
  \tabular{rl}{
    \code{River} \tab {One of four rivers: \code{Grasse}, \code{Oswegatchie}, \code{Raquette}, or \code{St. Regis}}\cr
    \code{Site} \tab {Location: 1=UpStream, 2=MidStream, 3=Downstream}\cr
    \code{Al} \tab {Aluminum}\cr
    \code{Ba} \tab {Barium}\cr
    \code{Br} \tab {Bromine}\cr
    \code{Ca} \tab {Calcium}\cr
    \code{Ce} \tab {Cerium}\cr
    \code{Cu} \tab {Copper}\cr
    \code{Dy} \tab {Dysprosium}\cr
    \code{Er} \tab {Erbim}\cr
    \code{Fe} \tab {Iron}\cr
    \code{Gd} \tab {Gadolinium}\cr
    \code{Ho} \tab {Holmum}\cr
    \code{K} \tab {Potassium}\cr
    \code{La} \tab {Lathanum}\cr
    \code{Li} \tab {Lithium}\cr
    \code{Mg} \tab {Magnesium}\cr
    \code{Mn} \tab {Manganese}\cr
    \code{Nd} \tab {Neodymium}\cr
    \code{Pr} \tab {Proseyodymium}\cr
    \code{Rb} \tab {Rubidium}\cr
    \code{Si} \tab {Silicon}\cr
    \code{Sr} \tab {Strontium}\cr
    \code{Y} \tab {Yttrium}\cr
    \code{Yb} \tab {Ytterbium}\cr
    \code{Zn} \tab {Zinc}\cr
    \code{Zr} \tab {Zirconium}\cr
  }
}
\details{
Some geologists were interested in the water chemistry of rivers in upstate New York. They took
water samples at three different locations in four rivers (Grasse, Oswegatchie, Raquette, and St.
Regis). The sampling sites were chosen to investigate how the composition of the water changes as
it flows from the source to the mouth of each river. The sampling sites were labeled as upstream,
midstream, and downstream. This dataset contains the concentrations (parts per million) of a variety of elements
in those water samples.
The dataset RiverIron contains the information for iron (FE) alone, along with the log of the concentration.
}
\source{
Thanks to Dr. Jeff Chiarenzelli of the St. Lawrence University Geology Department for the data.
}
\references{
Chiarenzelli, Lock, Cady, Bregani and Whitney,
"Variation in river multi-element chemistry related to bedrock buffering:
an example from the Adirondack region of northern New York, USA",
 Environmental Earth Sciences, Volume 67, Number 1 (2012), 189-204
}

\keyword{datasets}
