\name{RiverIron}
\alias{RiverIron}
\docType{data}
\title{Iron in River Water Samples}
\description{
Amounts of iron in water samples of four rivers
}
\format{
  A dataset with 12 observations on the following 4 variables.
  \tabular{rl}{
    \code{River} \tab {One of four rivers:  \code{Grasse}, \code{Oswegatchie}, \code{Raquette}, or \code{St. Regis}}\cr
    \code{Site} \tab {Location of the site: \code{DownStream}, \code{MidStream} or \code{Upstream}}\cr
    \code{Iron} \tab {Iron concentration in the water sample (parts per million)}\cr
    \code{LogIron} \tab {Log (base 10) of iron concentration}\cr
  }
}
\details{
Some geologists were interested in the water chemistry of rivers in upstate New York. They took
water samples at three different locations in four rivers (Grasse, Oswegatchie, Raquette, and St.
Regis). The sampling sites were chosen to investigate how the composition of the water changes as
it flows from the source to the mouth of each river. The sampling sites were labeled as upstream,
midstream, and downstream. This dataset contains the concentrations of iron in the samples.
The dataset RiverElements has similar concentration data for many other elements.
}
\source{
Thanks to Dr. Jeff Chiarenzelli of the St. Lawrence University Geology Department for the data.
}
\references{
Chiarenzelli, Lock, Cady, Bregani and Whitney,
"Variation in river multi-element chemistry related to bedrock buffering:
an example from the Adirondack region of northern New York, USA",
 Environmental Earth Sciences, Volume 67, Number 1 (2012), 189-204
}

\keyword{datasets}
