\name{SandwichAnts}
\alias{SandwichAnts}
\docType{data}
\title{Ants on Sandwiches}
\description{
Ant counts on samples of different kinds of sandwiches
}
\format{
  A dataset with 48 observations on the following 5 variables.
  \tabular{rl}{
    \code{Trial} \tab {Trial number}\cr
    \code{Bread} \tab {Type of bread:  \code{Multigrain}, \code{Rye}, \code{White}, or \code{Wholemeal}}\cr
    \code{Filling} \tab {Type of filling: \code{HamPickles}, \code{PeanutButter}, or \code{Vegemite}}\cr
    \code{Butter} \tab {Butter on the sandwich? \code{no}  or \code{yes}}\cr
    \code{Ants} \tab {Number of ants on the sandwich}\cr
  }
} \details{ As young students, Dominic Kelly and his friends enjoyed watching
ants gather on pieces of sandwiches.  Later, as a university student, Dominic
decided to study this with a more formal experiment. He chose three types of
sandwich fillings (vegemite, peanut butter, and ham & pickles),  four types
of bread (multigrain, rye, white, and wholemeal), and put butter on some of
the sandwiches. \cr
To conduct the experiment he randomly chose a sandwich,
broke off a piece, and left it on the ground near an ant hill.  After several
minutes he placed a jar over the sandwich bit and counted the number of
ants. He repeated the process, allowing time for ants to return to the hill
after each trial, until he had two samples for each combination of the three
factors.
}
\source{
Margaret Mackisack, ``Favourite Experiments: An Addendum to What is the Use of Experiments Conducted by Statistics Students?",
Journal of Statistics Education (1994)\cr
 http://www.amstat.org/publications/jse/v2n1/mackisack.supp.html
}

 \keyword{datasets}
