\name{SATGPA}
\alias{SATGPA}
\docType{data}
\title{SAT scores and GPA}
\description{
A sample of SAT scores and grade point averages for statistics students
}
%\useage{data(SATGPA)}
\format{
  A dataset with 24 observations on the following 3 variables.
  \tabular{rl}{
    \code{MathSAT} \tab {Score (out of 800) on the mathematics portion of the SAT exam}\cr
    \code{VerbalSAT} \tab {Score (out of 800) on the verbal portion of the SAT exam}\cr
    \code{GPA} \tab {Grade point average (0.0-4.0 scale)}\cr
  }
}
\details{
In recent years many colleges have re-examined the traditional role the scores on the Scholastic
Aptitude Tests (SAT's) play in making decisions on which students to admit. Do SAT scores really
help predict success in college? To investigate this question a group of 24 introductory statistics
students supplied the data in this dataset showing their score on the Verbal portion of the SAT
as well as their current grade point average (GPA) on a 0.0-4.0 scale.
}
\source{
Student survey in an introductory statistics course.
}

\keyword{datasets}
