\name{Sparrows}
\alias{Sparrows}
\docType{data}
\title{Sparrow Measurements}
\description{
Weight and wing length for a sample of Savannah sparrows
}
\format{
  A dataset with 116 observations on the following 3 variables.
  \tabular{rl}{
    \code{Treatment} \tab {Nest adjustment: \code{control}, \code{enlarged}, or \code{reduced}}\cr
    \code{Weight} \tab {Weight (in grams)}\cr
    \code{WingLength} \tab {Wing length (in mm)}\cr
  }
}
\details{
Priscilla Erickson from Kenyon College collected data on a stratified random sample of 116 Savannah sparrows
at Kent Island. Nests that were reduced, controlled (no change), or enlarged.
}
\source{
We thank Priscilla Erickson and Professor Robert Mauck from the Department of Biology at Kenyon College for
allowing us to use these data.
}

\keyword{datasets}
