\name{SpeciesArea}
\alias{SpeciesArea}
\docType{data}
\title{Land Area and Mammal Species}
\description{
Land area and number of mammal species for islands in Southeast Asia
}
%\useage{data(SpeciesArea)}
\format{
  A dataset with 14 observations on the following 5 variables.
  \tabular{rl}{
    \code{Name} \tab {Name of the island}\cr
    \code{Area} \tab {Area (in sq. km)}\cr
    \code{Species} \tab {Number of mammal species}\cr
    \code{logArea} \tab {Natural logarithm (base e) of Area}\cr
    \code{logSpecies} \tab {Natural logarithm (base e) of Species}\cr
  }
}
\details{
This dataset shows the number of mammal species and the
area for 13 islands in Southeast Asia. Biologists have speculated that the number of species is
related to the size of an island and would like to be able to predict the number of species given the
size of an island.
}
\source{
Heaney, Lawrence R. (1984) "Mammalian species richness on islands on the Sunda Shelf, Southeast Asia,"
Oecologia, 61:11 17.
}

\keyword{datasets}
