\name{Speed}
\alias{Speed}
\docType{data}
\title{Highway Fatality Rates (Yearly)}
\description{
Highway fatality rates 1987-2007
}
%\useage{data(Speed)}
\format{
  A dataset with 21 observations on the following 3 variables.
  \tabular{rl}{
    \code{Year} \tab {Year (1987-2007)}\cr
    \code{FatalityRate} \tab {Number of fatalities on interstate highways (per 100 million vehicle-miles)}\cr
    \code{StateControl} \tab {\code{0}=1987-1994 or \code{1}=1995-2007}\cr
  }
}
\details{
In 1987 the federal government allowed the speed
limit on interstate highways to be 65 mph in most areas. In 1995 federal restrictions were eliminated,
so that states assumed control of setting speed limits on interstate highways.  This data set compares
fatality rates for years before and after the states assumed control for highway speed limits.
}
\source{
Data from the National Highway Safety Administration website at\cr
http://www-fars.nhtsa.dot.gov/Main/index.aspx
}

\keyword{datasets}
