\name{SuicideChina}
\alias{SuicideChina}
\docType{data}
\title{Suicide Attempts in Shandong, China}
\description{
Data on serious suicide attempts in Shandong, China
}
\format{
  A data frame with 2571 observations on the following 11 variables.
  \describe{
    \item{\code{Person_ID}}{ID number}
    \item{\code{Hospitalised}}{Hospitalised? (\code{no} or \code{yes})}
    \item{\code{Died}}{Died? (\code{no} or \code{yes})}
    \item{\code{Urban}}{Urban area? (\code{no}, \code{unknown}, or \code{yes})}
    \item{\code{Year}}{Year (2009, 2010, or 2011)}
    \item{\code{Month}}{Month (\code{1}=Jan through \code{12}=December)}
    \item{\code{Sex}}{Sex (\code{female} or \code{male})}
    \item{\code{Age}}{Age (years)}
    \item{\code{Education}}{Education level  (\code{iliterate}, \code{primary}, \code{Secondary}, \code{Tertiary}, or \code{unknown})}
    \item{\code{Occupation}}{One of ten occupation categories}
    \item{\code{method}}{One of nine possible methods}
  }
}
\details{
Data from a study of serious suicide attempts over three years in a predominantly rural population in Shandong, China.
}
\source{
Sun J, Guo X, Zhang J, Wang M, Jia C, Xu A (2015) "Incidence and fatality of serious suicide attempts in a predominantly rural population in Shandong, China: a public health surveillance study," BMJ Open 5(2): e006762. https://doi.org/10.1136/bmjopen-2014-006762

Data downloaded via Dryad Digital Repository. https://doi.org/10.5061/dryad.r0v35
}
\keyword{datasets}
