\name{Swahili}
\alias{Swahili}
\docType{data}
\title{Attitudes Towards Swahili in Kenyan Schools}
\description{
Attitudes towards the Swahili language among Kenyan school children
}
%\useage{data(Swahili)}
\format{
  A dataset with 480 observations on the following 4 variables.
  \tabular{rl}{
    \code{Province} \tab {\code{NAIROBI} or \code{PWANI}}\cr
    \code{Sex} \tab {\code{female} or \code{male}}\cr
    \code{Attitude.Score} \tab {Score (out a possible 200 points) on a survey of attitude towards the Swahili language}\cr
    \code{School} \tab {Code for the school: \code{A} through \code{L}}\cr
  }
}
\details{
Hamisi Babusa, a Kenyan scholar, administered a survey to 480
students from Pwani and Nairobi provinces about their attitudes towards the Swahili language. In
addition, the students took an exam on Swahili. From each province, the students were from 6
schools (3 girls schools and 3 boys schools) with 40 students sampled at each school, so half of the
students from each province were males and the other half females. The survey instrument contained
40 statements about attitudes towards Swahili and students rated their level of agreement to
each. Of these questions, 30 were positive questions and the remaining 10 were negative questions.
On an individual question the most positive response would be assigned a value of 5 while the most
negative response would be assigned a value of 1. By summing (adding) the responses to each
question, we can find an overall Attitude Score for each student. The highest possible score would
be 200 (an individual who gave the most positive possible response to every question). The lowest
possible score would be 40 (an individual who gave the most negative response to every question).
}
\source{
Thanks to Dr. Babusi of Kenyatta University for sharing these data.
}

\keyword{datasets}
