\name{Tadpoles}
\alias{Tadpoles}
\docType{data}
\title{Effects of a Fungus on Tadpoles}
\description{
Comparing intestine lengths for tadpoles with and without exposure to Bd fugus
}
\format{
  A data frame with 27 observations on the following 4 variables.
  \describe{
    \item{\code{Treatment}}{Exposed to fungus (\code{Bd}=yes or \code{Control}=no)}
    \item{\code{Body}}{Length of body  (in mm)}
    \item{\code{GutLength}}{Length of intestine (in mm)}
    \item{\code{MouthpartDamage}}{Measure of damage to the mouth (e.g. missing teeth) }
  }
}
\details{
Biologists wondered whether tadpoles can adjust the relative length of their intestines if they are exposed to a fungus called Batrachochytrium dendrobatidis (Bd).
}
\source{
Venesky MD, Hanlon SM, Lynch K, Parris MJ, Rohr JR. (2013) "Optimal digestion theory does not predict the effect of pathogens on intestinal plasticity," Biol Lett 9: 20130038.
http://dx.doi.org/10.1098/rsbl.2013.0038
}
\keyword{datasets}
