\name{TechStocks}
\alias{TechStocks}
\docType{data}
\title{Daily Prices of Three Tech Stocks }
\description{
Daily closing prices of Apple, Google, and Microsoft stocks (12/1/2015 to 12/1/2017)
}
\format{
  A data frame with 504 observations on the following 5 variables.
  \describe{
    \item{\code{Date}}{Date (coded as \code{mm/dd/yyyy})}
    \item{\code{AAPL}}{Apple Inc. closing price}
    \item{\code{GOOG}}{Alphabet Inc. (Google) closing price}
    \item{\code{MSFT}}{Microsoft Corp. closing price}
    \item{\code{t}}{Time index (1 to 505)}
  }
}
\details{
Closing price of Apple (AAPL), Google/Alphabet (GOOG) and Microsoft (MSFT) stocks for each trading day in a two-year period from 12/1/2015 to 12/1/2017.
}
\source{
Data downloaded using the Quandl R package (12/2/2017)
}
\keyword{datasets}
