\name{TeenPregnancy}
\alias{TeenPregnancy}
\docType{data}
\title{State Teen Pregnancy Rates}
\description{
State teen pregnancy rates, Civil War participation, and church attendance.
}
\format{
  A data frame with 50 observations on the following 4 variables.
  \describe{
    \item{\code{State}}{State abbreviation}
    \item{\code{CivilWar}}{Role in Civil War (\code{B}=border, \code{C}=Confederate, \code{O}=other, or \code{U}=union)}
    \item{\code{Church}}{Percentage who attended church in previous week (from a state survey)}
    \item{\code{Teen}}{Number of pregnancies per 1000 teenage girls in state}
  }
}
\details{
State level data on teen pregnancies, church attendance, and role in the U.S. Civil War.
}
\source{
2010 teen pregnancy rate, per 1000 teenage women, per year.  Source:  Guttmacher Institute, via Tanya Lewis (5 May 2014) "Teen pregnancy rates by state," https://www.livescience.com
}
\keyword{datasets}
