\name{TextPrices}
\alias{TextPrices}
\docType{data}
\title{Textbook Prices}
\description{
Prices and number of pages for a sample of college textbooks
}
%\useage{data(TextPrices)}
\format{
  A dataset with 30 observations on the following 2 variables.
  \tabular{rl}{
    \code{Pages} \tab {Number of pages in the textbook}\cr
    \code{Price} \tab {Price of the textbook (in dollars)}\cr
  }
}
\details{
Two undergraduate students at Cal Poly - San Luis Obispo took a random sample of 30
textbooks from the campus bookstore in the fall of 2006. They recorded the price and number of
pages in each book, in order to investigate the question of whether number of pages can be used
to predict price.
}
\source{
Student project
}

\keyword{datasets}
