\name{ThreeCars2017}
\alias{ThreeCars2017}
\docType{data}
\title{Price, Age, and Mileage of Three Used Car Models}
\description{
Data from cars.com for a sample of three different models of used cars in 2017
}
\format{
  A data frame with 90 observations on the following 7 variables.
  \describe{
    \item{\code{CarType}}{Model (\code{Accord}, \code{Maxima}, or \code{Mazda6})}
    \item{\code{Age}}{Age of used car (years)}
    \item{\code{Price}}{Price (in thousands of dollars)}
    \item{\code{Mileage}}{Mileage (in thousands of miles)}
    \item{\code{Mazda6}}{Is the car a Mazda6? (\code{1}=yes or \code{0}=no)}
    \item{\code{Accord}}{Is the car an Accord? (\code{1}=yes or \code{0}=no)}
    \item{\code{Maxima}}{Is the car a Maxima? (\code{1}=yes or \code{0}=no)}
  }
}
\details{
Data for a sample of cars from three models (Mazda6, Honda Accord, Nissan Maxima) from a website. The dataset AccordPrice is a subset of this file.
}
\source{
Data obtained from cars.com, February 2017 using zip code 44107, Lakewood, Ohio.
}
\keyword{datasets}
