\name{TipJoke}
\alias{TipJoke}
\docType{data}
\title{Tip Joke}
\description{
Effect of a waiter leaving a joke or an advertisement on getting a tip
}
%\useage{data(TipJoke)}
\format{
  A dataset with 211 observations on the following 5 variables.
  \tabular{rl}{
    \code{Card} \tab {Type of card used: \code{Ad}, \code{Joke}, or \code{None}}\cr
    \code{Tip} \tab {\code{1}=customer left a tip or \code{0}=no tip}\cr
    \code{Ad} \tab {Indicator for \code{Ad} card}\cr
    \code{Joke} \tab {Indicator for \code{Joke} card}\cr
    \code{None} \tab {Indicator for no card}\cr
  }
}
\details{
Can telling a joke affect whether or not a waiter in a coffee bar receives a tip from a customer?
A study investigated this question at a coffee bar at a famous resort on the west coast of France.
The waiter randomly assigned coffee-ordering customers to one of three groups: When receiving
the bill one group also received a card telling a joke, another group received a card containing an
advertisement for a local restaurant, and a third group received no card at all. He recorded whether or not
each customer left a tip.
}
\source{
Gueguen, Nicholas (2002), "The Effects of a Joke on Tipping When it is Delivered at the Same Time as the Bill," 
Journal of Applied Social Psychology, 32, 1955-1963.
}

\keyword{datasets}
