\name{TMS}
\alias{TMS}
\docType{data}
\title{Migraines and TMS}
\description{
Effects of transcranial magnetic stimulation (TMS) on migraine headaches
}
\format{
  A dataset with 2 observations on the following 4 variables.
  \tabular{rl}{
    \code{Group}\tab {Treatment group: \code{Placebo} or \code{TMS}}\cr
    \code{Yes}\tab {Count of number of patients that were pain-free}\cr
    \code{No}\tab {Count of number of patients that had pain}\cr
    \code{Trials}\tab {Number of patients in the group}\cr
  }
}
\details{
A study investigated whether a handheld device that sends a magnetic pulse into a person's head
might be an effective treatment for migraine headaches. Researchers recruited 200 subjects who
suffered from migraines and randomly assigned them to receive either the TMS (transcranial magnetic
stimulation) treatment or a sham (placebo) treatment from a device that did not deliver any
stimulation. Subjects were instructed to apply the device at the onset of migraine symptoms and
then assess how they felt two hours later.  This dataset is a two-way table of the results.

This dataset renamed as Migraines in second edition.
}
\source{
Based on results in R. B. Lipton, et. al. (2010) ``Single-pulse Transcranial Magnetic Stimulation
for Acute Treatment of Migraine with Aura: A Randomised, Double-blind, Parallel-group,
Sham-controlled Trial," 9(4):373-380.
}

\keyword{datasets}
