\name{TomlinsonRush}
\alias{TomlinsonRush}
\docType{data}
\title{
LaDainian Tomlinson Rushing Yards
}
\description{
Rushing yards for each game LaDainian Tomlinson played in the 2006 National Football League (NFL regular) season.
}
\format{
  A dataset with 16 observations on the following 4 variables.
  \tabular{rl}{
    \code{Game}\tab {Week number in the 2006 season}\cr
    \code{Opponent}\tab {Name of opposing team}\cr
    \code{Attempts}\tab {Number of rushing attempts}\cr
    \code{Yards}\tab {Total yards gained rushing for the game}\cr
  }
}
\details{
For each of the sixteen games the San Diego Chargers played in the 2006 NFL regular season
we have the number of times LaDainian Tomlinson ran the ball and the total yards he gained.

This data set from the first edition was replaced by BreesPass in the second edition.
}
\source{
Data downloaded from http://www.pro-football-reference.com/players/T/TomlLa00/gamelog/2006/
}

\keyword{datasets}
