% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyNonaddPlot.R
\name{TukeyNonaddPlot}
\alias{TukeyNonaddPlot}
\title{Tukey Nonadditivity Plot for Two-way ANOVA}
\usage{
TukeyNonaddPlot(formula, data, out = "n",
  main = "Tukey Nonadditivity Plot", ylab = "Residuals")
}
\arguments{
\item{formula}{A formula for a two-way ANOVA in the form Response=FactorA+FactorB (or FactorA*FactorB)}

\item{data}{A dataframe}

\item{out}{Control what is returned. Default is "n"=nothing. Other otions are "comp" for the comparisons, "line" for the equaotn of the line, and "resid" for the cell residuals}

\item{main}{Add a title, default is "Tukey Nonadditivity Plot"}

\item{ylab}{Label vertical axis, default is "Resiuduals"}
}
\value{
Depends on optoin set with \code{out}.
}
\description{
This function produces a Tukeuy Nonadditivity plot for a two-way ANOVA model.
}
\details{
More details need to be written
}
\examples{
require(mosaic)
data(Dinosaurs)
TukeyNonaddPlot(Iridium~Source*factor(Depth),data=Dinosaurs)
}
