\name{TwinsLungs}
\alias{TwinsLungs}
\docType{data}
\title{Comparing Twins Ability to Clear Radioactive Particles}
\description{
Experiment comparing twins (one urban, one rural) ability to clear airborne radioactive particles from their lungs
}
\format{
  A dataset with 14 observations on the following 3 variables.
  \tabular{rl}{
    \code{Pair}\tab {Code for the twin pair: \code{A} - \code{G}}\cr
    \code{Environ}\tab {Living environment:  \code{Rural} or \code{Urban}}\cr
    \code{Percent}\tab {Percentage of radioactivity remaining in lungs}\cr
  }
}
\details{
This dataset is from a study to compare the effect of living environment (rural or urban) on human lung function,
where the researchers were able to locate seven pairs of twins with one twin in each pair living
in the country, the other in a city. To measure lung function, twins inhaled  an aerosol of radioactive Teflon particles.   By
measuring the  level of radioactivity immediately  and  then  again  after  an hour,  the  scientists  could
measure the  rate  of ``tracheobronchial clearance."   The  percentage  of radioactivity remaining
in the  lungs after  an hour  told  how quickly subjects'  lungs cleared the  inhaled  particles.

This dataset was renamed as RadioactiveTwins for the second edition.

}
\source{
``Urban factor and tracheobronchial clearance" by Per Camner and Klas Philipson in Archives of Environmental
Health, V. 27 (1973), page 82. Data can be found in Introduction to Mathematical Statistics and its Applications,
2nd Edition by Richard J. Larson and Morris L. Marx. Englewood Cliffs, NJ: Prentice Hall, p. 548.
}

\keyword{datasets}
