\name{USstamps}
\alias{USstamps}
\docType{data}
\title{Price of US Stamps}
\description{
Price of US stamp for first class mail 1885-2012
}
\format{
  A dataset with 25 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year}\tab {Years when stamp price changed}\cr
    \code{Price}\tab {Cost of a US first class stamp (in cents)}\cr
 }
}
\details{
The data record the year and price for each change in price for a
US first class (1 ounce, domestic letter) stamp since 1885.
}
\source{
http://about.usps.com/who-we-are/postal-history/domestic-letter-rates-1863-2011.htm
}

\keyword{datasets}
