\name{VisualVerbal}
\alias{VisualVerbal}
\docType{data}
\title{Visual versus Verbal Performance}
\description{
Experiment to compare visual and verbal performance
}
\format{
  A data frame with 80 observations on the following 5 variables.
  \describe{
    \item{\code{Subject}}{Subject number (\code{s1} to \code{s20})}
    \item{\code{Task}}{Follow a letter (\code{Visual}) or a sentence (\code{Verbal})}
    \item{\code{Report}}{Point response (\code{Visual}) or say response (\code{Verbal})}
    \item{\code{Group}}{Combination of Task+Report (\code{Letter Point}, \code{Letter Say}, \code{Sentence Point}, or \code{Sentence Say})}
    \item{\code{Time}}{Response time (in seconds)}
  }
}
\details{
Subjects carried out two kinds of tasks, one visual (identify letters), one verbal (identify sentences); and to report the results in either of two ways, one visual (pointing at a response), one verbal (speaking a response).  Time to complete each task was recorded in seconds.
}
\source{
Original experiment from Brooks, L., R. (1968)  "Spatial and verbal components of the act of recall," Canadian J. Psych. V 22, pp. 349 - 368.  These data collected from a Mount Holyoke College psychology class.
}
\keyword{datasets}
