\name{Volts}
\alias{Volts}
\docType{data}
\title{Volts}
\description{
Voltage drop over time as a capacitor discharges
}
%\usage{data(Volts)}
\format{
  A dataset with 50 observations on the following 2 variables.
  \tabular{rl}{
    \code{Voltage}\tab {Voltage (in volts)}\cr
    \code{Time}\tab {Time after charging (in seconds)}\cr
  }
}
\details{
A capacitor was charged with a nine-volt battery and then a voltmeter
recorded the voltage as the capacitor was discharged. Measurements were taken every .02 seconds.
}
\source{
Measurements recorded by one of the authors. 
}

\keyword{datasets}
