\name{WalkingBabies}
\alias{WalkingBabies}
\docType{data}
\title{Effects of Exercise on First Walking}
\description{
An experiment to see if special exercises help babies learn to walk sooner
}
\format{
  A dataset with 24 observations on the following 2 variables.
  \tabular{rl}{
    \code{Group}\tab {Treatments: \code{exercise control}, \code{final report}, \code{special exercises}, or \code{weekly report}}\cr
    \code{Age}\tab {Age (in months) when first walking}\cr
  }
}
\details{
Scientists wondered if they could get babies to walk sooner by prescribing a set of special exercises.
Their experimental design  included four groups of babies and the following treatments:\cr

Special exercises: Parents were shown the special exercises and encouraged to use them with
their children. They were phoned weekly to check on their child's progress.\cr

Exercise control: These parents were not shown the special exercises, but they were told to
make sure their babies spent at least 15 minutes a day exercising. \cr

Weekly report: Parents in this group were not given instructions about exercise. Like the
parents in the treatment group, however, they received a phone call each week to check on
progress.\cr

Final report: These parents were not given weekly phone calls or instructions about exercises.
They reported at the end of the study.
}
\source{
Zelazo, Phillip R., Nancy Ann Zelazo, and Sarah Kolb (1972), ``Walking in the Newborn," Science, v. 176, pp.
314-315.
}

\keyword{datasets}
