\name{WalkTheDogs}
\alias{WalkTheDogs}
\docType{data}
\title{Did the Author Walk the Dogs Today?}
\description{
Daily pedometer data for one of the authors
}
\format{
  A data frame with 223 observations on the following 7 variables.
  \describe{
    \item{\code{StepCount}}{Number of steps taken in the day}
    \item{\code{Kcal}}{Calories burned (according to pedometer)}
    \item{\code{Miles}}{Miles walked}
    \item{\code{Weather}}{\code{cold}, \code{rain}, or \code{shine}}
    \item{\code{Day}}{Day of week (\code{F}=Friday, \code{M}=Monday, \code{R}=Thursday, \code{S}=Saturday, \code{T}=Tuesday, \code{U}=Sunday, \code{W}=Wednesday)}
    \item{\code{Walk}}{Were the dogs walked?  (\code{1}=yes or \code{0}=no)}
    \item{\code{Steps}}{Steps in units of 1,000 (so StepCount/1000)}
  }
}
\details{
One of the authors recorded daily pedometer data, the weather, and whether or not he walked the dogs.
}
\source{
One of the author's pedometer records.
}
\keyword{datasets}
