\name{Whickham2}
\alias{Whickham2}
\docType{data}
\title{Whickham Health Study}
\description{
Morality data over 20 years for 1314 women from Whickham, England
}
\format{
  A data frame with 1314 observations on the following 5 variables.
  \describe{
    \item{\code{Outcome}}{Status at 20-year follow-up (\code{Alive} or \code{Dead})}
    \item{\code{Smoker}}{Smoker at baseline? (\code{No} or \code{Yes})}
    \item{\code{Age}}{Age (in years at baseline)}
    \item{\code{AgeGroup}}{Age group (\code{18-64} or \code{65+})}
    \item{\code{Alive}}{Numeric code for Outcome (\code{1}=alive or \code{0}=dead)}
  }
}
\details{
Twenty-year mortality, smoking status, and age for 1314 women in Whickham, England.  We have named this Whickham2 to distinguish it from Whickham, which is a file in the mosaicData package.
}
\source{
A version of these data are in the mosaicData package but originally are from:

Appleton, D. R., French, J. M., and Vanderpump, M.P. (1996), "Ignoring a Covariate: An Example of Simpson's Paradox," The American Statistician, 50, 340-341.
}
\keyword{datasets}
