\name{WordMemory}
\alias{WordMemory}
\docType{data}
\title{WordMemory}
\description{
Percentage of different types of words recalled
}
%\usage{data(WordMemory)}
\format{
  A dataset with 40 observations on the following 4 variables.
  \tabular{rl}{
    \code{Subject}\tab {Code to identify each subject: \code{A} to \code{J}}\cr
    \code{Abstract}\tab {Words were abstract? \code{No} or \code{Yes}}\cr
    \code{Frequent}\tab {Words were common? \code{No} or \code{Yes}}\cr
    \code{Percent}\tab {Percentage of words recalled (out of 25)}\cr
  }
}
\details{
One hundred words were presented to each subject in a randomized order.
The goal of the experiment was to see whether some kinds of words were easier to remember than
others. In particular, are common words like potato, love, diet, and magazine easier to remember
than less common words like manatee, hangnail, fillip, and apostasy? Are concrete words like coffee,
dog, kale, and tamborine easier than abstract words like beauty, sympathy, fauna, and guile? There
were 25 words each of four kinds, obtained by crossing the two factors of interest, Abstraction
(concrete or abstract) and Frequency (common or rare).
}
\source{
Data from a student laboratory project, Department of Psychology and Education, Mount Holyoke College.
}

\keyword{datasets}
