\name{YouthRisk}
\alias{YouthRisk}
\docType{data}
\title{Annual survey of health-risk youth behaviors}
\description{
Data from the Youth Risk Behavior Surveillance System
}
\format{
  A data frame with 13387 observations on the following 6 variables.
  \describe{
    \item{\code{ride.alc.driver}}{\code{1}=rode with a drinking driver in past 30 days or \code{0}=did not}
    \item{\code{female}}{\code{1}=female or \code{0}=male}
    \item{\code{grade}}{Year in high school: \code{9}, \code{10}, \code{11}, or \code{12}}
    \item{\code{age4}}{Age (in years)}
    \item{\code{smoke}}{Ever smoked? \code{1}=yes or \code{0}=no}
    \item{\code{DriverLicense}}{Have a driver's license? \code{1}=yes or \code{0}=no}
  }
}
\details{
This dataset is derived from
the 2007 Youth Risk Behavior Surveillance System (YRBSS), which is an annual survey conducted
by the Centers for Disease Control and Prevention (CDC) to monitor the prevalence of health-risk
youth behaviors. This dataset focuses on whether or not youths have recently (in past 30  days)
ridden with a drunk driver.
}
\source{
http://www.cdc.gov/HealthyYouth/yrbs/index.htm
}
\keyword{datasets}
