\name{YouthRisk2007}
\alias{YouthRisk2007}
\docType{data}
\title{Riding with a Driver Who Has Been Drinking}
\description{
Risky behavior (riding with a drinking driver) in youths
}
\format{
  A dataset with 13387 observations on the following 6 variables.
  \tabular{rl}{
    \code{ride.alc.driver}\tab {\code{1}=rode with a drinking driver in past 30 days or \code{0}=did not}\cr
    \code{female}\tab {\code{1}=female or \code{0}=male}\cr
    \code{grade}\tab {Year in high school: \code{9}, \code{10}, \code{11}, or \code{12}}\cr
    \code{age4}\tab {Age (in years)}\cr
    \code{smoke}\tab {Ever smoked? \code{1}=yes or \code{0}=no}\cr
    \code{DriverLicense}\tab {Have a driver's license? \code{1}=yes or \code{0}=no}\cr
  }
}
\details{
This dataset is derived from
the 2007 Youth Risk Behavior Surveillance System (YRBSS), which is an annual survey conducted
by the Centers for Disease Control and Prevention (CDC) to monitor the prevalence of health-risk
youth behaviors. This dataset focuses on whether or not youths have recently (in past 30  days)
ridden with a drunk driver.

This dataset renamed as YouthRisk for the second edition.
}
\source{
The article "Which Young People Accept a Lift From a Drunk or Drugged Driver?" in Accident Analysis and
Prevention (July 2009. pp. 703-9) provides more details.
}
\references{
A more recent version of the full dataset is available at http://www.cdc.gov/brfss/technical_infodata/surveydata.htm.
}
\keyword{datasets}
