\name{YouthRisk2009}
\alias{YouthRisk2009}
\docType{data}
\title{YouthRisk2009}
\description{
Survey of students in grades 9-12 concerning health-risk behiviors
}
%\usage{data(YouthRisk2009)}
\format{
  A dataset with 500 observations on the following 6 variables.
  \tabular{rl}{
    \code{Sleep}\tab {Average hours sleep on school night, coded with \code{10 or more hours}, \code{9 hours},} \cr
 ]tab {\code{8 hours}, \code{7 hours}, \code{6 hours}, \code{5 hours}, or \code{4 or less hours}}\cr
    \code{Sleep7}\tab {Seven or more hours of sleep? 0=\code{No}, 1=\code{Yes}}\cr
    \code{SmokeLife}\tab {Ever smoked? \code{No} or \code{Yes}}\cr
    \code{SmokeDaily}\tab {Regular smoker? \code{No} or \code{Yes}} \cr
    \code{MarijuaEver}\tab {Ever smoked marijuana? 0=\code{No} or 1=\code{Yes}}\cr
    \code{Age}\tab {Age (in years)} \cr
   }
}
\details{
Data from the Centers for Disease Control's Youth Risk Behavior Surveillance System (YRBSS). 
}
\source{
http://www.cdc.gov/HealthyYouth/yrbs/index.htm
}

\keyword{datasets}
