\name{AccordPrice}
\alias{AccordPrice}
\docType{data}
\title{Prices of Used Honda Accords (in 2017)}
\description{
Age, price, and mileage of used Honda Accords in 2017
}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{Age}}{Age of used Honda Accord car}
    \item{\code{Price}}{Price (in $1,000's)}
    \item{\code{Mileage}}{Mileage (in 1,000's of miles)}
  }
}
\details{
Information on used Honda Accords obtained from cars.com.
}
\source{
Cars.com, February 2017 using zip code 44107, Lakewood, Ohio
}
\keyword{datasets}
