\name{AlcoholResponse}
\alias{AlcoholResponse}
\docType{data}
\title{
AlcoholResponse
}
\description{
Reaction times in a driving simulation after drinking alcohol or a placebo.
}
%\useage{data(AlcoholResponse)}
\format{
  A dataset with 20 observations on the following 2 variables.
  \tabular{rl}{
    \code{Response} \tab {Reaction time (seconds)}\cr
    \code{Treatment} \tab {Type of drink: \code{Placebo} or \code{Alcohol}}\cr
  }
}
\details{
Participants were randomly assigned to consume either an alcoholic or nonalcoholic (placebo) drink
before their mean reaction time was measured for a series of simulated driving situations. 
}
\source{
??
}

\keyword{datasets}
