\name{AlcoholResponse2}
\alias{AlcoholResponse2}
\docType{data}
\title{
AlcoholResponse2
}
\description{
Reaction times in a driving simulation after drinking alcohol or a placebo.
}
\usage{data(AlcoholResponse2)}
\format{
  A data frame with 10 observations on each of the following 2 variables.
  \describe{
    \item{\code{Placebo}}{Reaction time for participants in the placebo group}
    \item{\code{Alcohol}}{Reaction time for participants in the alcohol group}
  }
}
\details{
Participants were randomly assigned to consume either an alcoholic or nonalcoholic (placebo) drink
before their mean reaction time was measured for a series of simulated driving situations. 
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(AlcoholResponse2)
}
\keyword{datasets}
