\name{AthleteGrad}
\alias{AthleteGrad}
\docType{data}
\title{Athletic Participation, Race, and Graduation}
\description{
Six-year graduation data for 214,555 students in 2004
}
\format{
  A data frame with 214555 observations on the following 3 variables.
  \describe{
    \item{\code{Student}}{\code{Athlete} or \code{NonAthlete}}
    \item{\code{Race}}{ \code{Black} or  \code{White}}
    \item{\code{Grad}}{\code{1}=graduated within 6 years, otherwise \code{0}}
  }
}
\details{
Six-year graduation data from 2004 for male non-athletes and for male athletes, where "Athlete"" means football or basketball player. These data show Simpson's Paradox.
}
\source{
Victor Matheson, College of the Holy Cross, collected the summary statistics.

Data are derived from the summary tables in:

Matheson, V., "Research Note: Athletic Graduation Rates and Simpson's Paradox," Economics of Education Review, Vol. 26:4 (August 2007), 516-520.


}
\keyword{datasets}
