\name{BirdNest}
\alias{BirdNest}
\docType{data}
\title{
Nest Characteristics for Different Bird Species
}
\description{
Nest and species characteristics for North American passerines
}
\format{
  A data frame with 84 observations on the following 12 variables.
  \tabular{rl}{
    \code{Species} \tab {Latin species name}\cr
    \code{Common} \tab {Common species name}\cr
    \code{Page} \tab {Page in a bird manual describing the species}\cr
    \code{Length} \tab {Mean body length for the species (in cm)}\cr
    \code{Nesttype} \tab {Type of nest}\cr
    \code{Location} \tab {Location of nest}\cr
    \code{No.eggs} \tab {Number of eggs}\cr
    \code{Color} \tab {Egg color (\code{0}=plain/solid or \code{1}=speckled/spotted)}\cr
    \code{Incubate} \tab {Mean length of time (in days) the species incubates eggs in the nest}\cr
    \code{Nestling} \tab {Mean length of time (in days) the species cares for babies in the nest until fledged}\cr
    \code{Totcare} \tab {Total care time = Incubate+Nestling}\cr
    \code{Closed} \tab {1=closed nest (pendant, spherical, cavity, crevice, burrow), 0=open nest (saucer, cup)}\cr
  }
}
\details{
Amy R. Moore, as a student at Grinnell College in 1999, wanted to study the relationship between
species characteristics and the type of nest a bird builds, using data collected from available
sources. For the study, she collected data by species for 84 separate species of North American
passerines.
}
\source{
Project by Amy Moore at Grinnell College
}
\references{
The Birders Handbook, by Ehrlich, et al. (1988)
}

\keyword{datasets}
