\name{Cereal}
\alias{Cereal}
\docType{data}
\title{Cereal}
\description{
Breakfast cereals
}
%\useage{data(Cereal)}
\format{
  A dataset with 36 observations on the following 4 variables.
  \tabular{rl}{
    \code{Cereal} \tab {Brandname of cereal}\cr
    \code{Calories} \tab {Calories per serving}\cr
    \code{Sugar} \tab {Grams of sugar per serving}\cr
    \code{Fiber} \tab {Grams of fiber per serving}\cr
  }
}
\details{
Data give nutrition contents (per serving) for 36 breakfast cereals.
} 
\source{
Data collected from nutrition labels. 
}

\keyword{datasets}
