\name{Cuckoo}
\alias{Cuckoo}
\docType{data}
\title{
Cuckoo
}
\description{
Lengths of cuckoo eggs laid in other birds' nests
}
\usage{data(Cuckoo)}
\format{
  A data frame with 120 observations on the following 2 variables.
  \describe{
    \item{\code{Bird}}{Type of bird nest: \code{mdw_pipit} (meadow pipit), 
\code{tree_pipit}, \code{hedge_sparrow}, \code{robin}, \code{wagtail}, \code{wren}}
    \item{\code{Length}}{Cuckoo egg length (in mm)}
  }
}
\details{
Cuckoos are knows to lay their eggs in the nests of other (host) birds. 
The eggs are then adopted and hatched by the host birds.  
The data give the lengths of cuckoo eggs found in nests of varous other bird species.
}
\source{
Downloaded from DASL at http://lib.stat.cmu.edu/DASL/Datafiles/cuckoodat.html
}
\references{
L.H.C. Tippett, The Methods of Statistics, 4th Edition, John Wiley and Sons, Inc., 1952, p. 176. 
}
\examples{
%% data(Cuckoo)
}
\keyword{datasets}
