\name{Election08}
\alias{Election08}
\docType{data}
\title{Election08}
\description{
State-by-state information from the 2008 U.S. presidential election
}
%\useage{data(Election08)}
\format{
  A data frame with 51 observations on the following 7 variables.
  \tabular{rl}{
    \code{State} \tab {Name of the state}\cr
    \code{Abr} \tab {Abbreviation for the state}\cr
    \code{Income} \tab {Per capita income in the state as of 2007 (in dollars)}\cr
    \code{HS} \tab {Percentage of adults with at least a high school education}\cr
    \code{BA} \tab {Percentage of adults with at least a college education}\cr
    \code{Dem.Rep} \tab {Difference in \%Democrat-\%Republican (according to 2008 Gallup survey)}\cr
    \code{ObamaWin} \tab {\code{1}= Obama (Democrat) wins state in 2008 or \code{0}=McCain (Republican wins)}\cr
  }
}
\details{
This dataset contains information from all 50 states and the District of Columbia for the 2008 U.S. presidential election. 
}
\source{
Various online sources. 
}

\keyword{datasets}
