\name{Hawks}
\alias{Hawks}
\docType{data}
\title{Hawks}
\description{
Data for a sample of hawks
}
%\useage{data(Hawks)}
\format{
  A dataset with 908 observations on the following 19 variables.
  \tabular{rl}{
    \code{Month} \tab {code{8}=September to \code{12}=December}\cr
    \code{Day} \tab {Date in the month}\cr
    \code{Year} \tab {Year: 1992-2003}\cr
    \code{CaptureTime} \tab {Time of capture (HH:MM)}\cr
    \code{ReleaseTime} \tab {Time of release (HH:MM)}\cr
    \code{BandNumber} \tab {ID band code}\cr
    \code{Species} \tab {\code{CH}=Cooper's, \code{RT}=Red-tailed,  \code{SS}=Sharp-Shinned}\cr
    \code{Age} \tab {\code{A}=Adult or \code{I}=Imature}\cr
    \code{Sex} \tab {\code{F}=Female or \code{M}=Male}\cr
    \code{Wing} \tab {a numeric vector}\cr
    \code{Weight} \tab {a numeric vector}\cr
    \code{Culmen} \tab {Size of beak}\cr
    \code{Hallux} \tab {a numeric vector}\cr
    \code{Tail} \tab {Tail length (in cm)}\cr
    \code{StandardTail} \tab {a numeric vector}\cr
    \code{Tarsus} \tab {a numeric vector}\cr
    \code{WingPitFat} \tab {a numeric vector}\cr
    \code{KeelFat} \tab {a numeric vector}\cr
    \code{Crop} \tab {a numeric vector}\cr
  }
}
\details{
Students and faculty at Cornell College in Mount Vernon, Iowa, collected the data over many years 
at the hawk blind at Lake MacBride near Iowa City, Iowa. The data set that we are analyzing here
 is a subset of the original data set, using only those species for which there were more than 10 observations. 
 Data were collected on random samples of three different species of hawks: 
 Red-tailed, Sharp-shinned, and Cooper's hawks. 
}
\source{
Many thanks to the late Professor Bob Black at Cornell College for sharing these data with us.
}

\keyword{datasets}
