\name{HawkTail}
\alias{HawkTail}
\docType{data}
\title{
HawkTail
}
\description{
Tail lengths for two hawk species
}
\usage{data(HawkTail)}
\format{
  A data frame with 838 observations on the following 2 variables.
  \describe{
    \item{\code{Species}}{\code{RT}=Red-tailed, \code{SS}=Sharp-shinned}
    \item{\code{Tail}}{Length of tail (in mm)}
  }
}
\details{
Tail lengths measured for a sample of 838 hawks observed in Mount Vernon, Iowa. 
Note: HawkTail2 has similar data in unstacked format for three species of hawks.
}
\source{
Observations by students and faculty at Cornell College
}
\references{
%%
}
\examples{
%% data(HawkTail)
}
\keyword{datasets}
