\name{HawkTail}
\alias{HawkTail}
\docType{data}
\title{Tail Lengths of Hawks}
\description{
Tail lengths for two hawk species
}
\format{
  A data frame with 838 observations on the following 2 variables.
  \tabular{rl}{
    \code{Species} \tab {\code{RT}=Red-tailed, \code{SS}=Sharp-shinned}\cr
    \code{Tail} \tab {Length of tail (in mm)}\cr
  }
}
\details{
Tail lengths measured for a sample of 838 hawks observed in Mount Vernon, Iowa.
Note: HawkTail2 has these data in unstacked format and they are a subset of the data in Hawks which has a third species (Cooper's hawk).
}
\source{
Observations by students and faculty at Cornell College.
}

\keyword{datasets}
