\name{HawkTail2}
\alias{HawkTail2}
\docType{data}
\title{
HawkTail2
}
\description{
Tail lengths for three hawk species
}
\usage{data(HawkTail2)}
\format{
  A data frame with observations on the following 3 variables.
  \describe{
    \item{\code{Tail_CH}}{Tail length (in mm) for a sample of Cooper's hawks}
    \item{\code{Tail_RT}}{Tail length (in mm) for a sample of Red-tailed hawks}
    \item{\code{Tail_SS}}{Tail length (in mm) for a sample of Sharp-shinned hawks}
  }
}
\details{
Tail lengths measured for a sample of 908 hawks observed in Mount Vernon, Iowa. 
Note: HawkTail has similar data in stacked format for just the Red-tailed and Sharp-shinned hawks.
}
\source{
Observations by students and faculty at Cornell College.
}
\references{
%%
}
\examples{
data(HawkTail2)
}
\keyword{datasets}
