\name{HawkTail2}
\alias{HawkTail2}
\docType{data}
\title{Tail Lengths of Hawks (Unstacked)}
\description{
Tail lengths for two hawk species
}
\format{
  A data frame with observations on the following 2 variables.
  \tabular{rl}{
    \code{Tail_RT} \tab {Tail length (in mm) for a sample of Red-tailed hawks}\cr
    \code{Tail_SS} \tab {Tail length (in mm) for a sample of Sharp-shinned hawks}\cr
  }
}
\details{
Tail lengths measured for a sample of hawks observed in Mount Vernon, Iowa.
Note: HawkTail has similar data in stacked format. The Hawks dataset has more variables and a third species (Cooper's hawk).
}
\source{
Observations by students and faculty at Cornell College.
}

\keyword{datasets}
