\name{InfantMortality2010}
\alias{InfantMortality2010}
\docType{data}
\title{Infant Mortality Rates}
\description{
Infant mortality rates in the United States by decade (1920-2010)
}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{Mortality}}{Deaths within one year of birth (per 1000 births)}
    \item{\code{Year}}{Year (1920-2010 by decades)}
  }
}
\details{
Infant mortality (deaths within one year of birth per 1,000 births) in the US from 1920 - 2010 (by decade).
}
\source{
CDC National Vital Statistics Reports at http://www.cdc.gov/nchs/data/nvsr/nvsr57/nvsr57_14.pdf and
 https://www.cdc.gov/nchs/data/nvsr/nvsr64/nvsr64_09.pdf
}

\keyword{datasets}
