\name{LongJumpOlympics}
\alias{LongJumpOlympics}
\docType{data}
\title{Olympic Men's Long Jump Gold Medal Distance (1900 - 2008)}
\description{
Winning distances in men's Olympic long jump competitions (1900 - 2008)
}
\format{
  A data frame with 26 observations on the following 2 variables.
  \tabular{rl}{
    \code{Year} \tab {Year of the Olympics (1900 - 2008)}\cr
    \code{Gold} \tab {Winning men's long jump distance (in meters)}\cr
  }
}
\details{
Gold medal winning distances for the men's long jump at the Olympics from 1900 to 2008.

Updated to LongJumpOlympics2016 in second edition.
}
\source{
Historical Olympic long ump results at http://trackandfield.about.com/od/longjump/qt/olymlongjumpmen.htm
}

\keyword{datasets}
