\name{LosingSleep}
\alias{LosingSleep}
\docType{data}
\title{Sleep Hours for Teenagers}
\description{
Hours of sleep for teenagers
}
\format{
  A data frame with 446 observations on the following 3 variables.
  \describe{
    \item{\code{Person}}{Cased ID number}
    \item{\code{Age}}{Age (in years)}
    \item{\code{Outcome}}{Average at least 7 hours of sleep? (\code{1=yes} or \code{0}=no)}
  }
}
\details{
Data from a sample of 446 teens, aged 14 to 18, who answer the question, "On an average school night, how many hours of sleep do you get?" The outcome variable records whether or not each person averages at least 7 hours of sleep.
}
\source{
Wahlstrom, K., Dretzke, B., Gordon, M., Peterson, K., Edwards, K., & Gdula, J. (2014)
"Examining the Impact of Later School Start Times on the Health and Academic Performance of
High School Students: A Multi-Site Study,"  Center for Applied Research and Educational
Improvement. St Paul, MN:  University of Minnesota.
}
\keyword{datasets}
