\name{RadioactiveTwins}
\alias{RadioactiveTwins}
\docType{data}
\title{Comparing Twins Ability to Clear Radioactive Particles}
\description{
Experiment comparing twins (one urban, one rural) ability to clear airborne radioactive particles from their lungs
}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{TwinPair}}{Identifies the twin pairs (1 to 15)}
    \item{\code{Env}}{Residential environment  (\code{Rural} or \code{Urban})}
    \item{\code{Rate}}{Clearance rate (percentage radioactive particles remaining after one hour)}
  }
}
\details{
To assess lung health, the scientists measured "tracheobronchial clearance rate," that is, in English, "How fast do your lungs get rid of nasty stuff?"  Each subject agreed to inhale an aerosol of radioactive Teflon particles.  A Geiger counter held to the chest measured the radioactivity just after inhaling, and again one hour later.  The clearance rate was the percentage of radioactivity remaining -- the lower the better.  Subjects were 15 sets of identical twins, each pair with one twin living in an urban environment and the other in a rural environment.
}
\source{
Per Camner MD & Klas Philipson MSc (1973) "Urban Factor and Tracheobronchial Clearance," Archives of Environmental Health: An International Journal, 27:2, 81-84, DOI: 10.1080/00039896.1973.10666323
Link to the article: https://doi.org/10.1080/00039896.1973.10666323
}
\keyword{datasets}
