\name{SleepingShrews}
\alias{SleepingShrews}
\docType{data}
\title{Shrew Heart Rates at Stages of Sleep}
\description{
Heart rates for a sample of six tree shrews at each of three stages of sleep.
}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{Row ID}
    \item{\code{Shrew}}{Shrew ID (\code{A} through \code{F})}
    \item{\code{Phase}}{Phase of sleep  (\code{DSW}=deep wave, \code{LSW}=light wave, or \code{REM}=dreaming)}
    \item{\code{Rate}}{Heart rate (beats per minute)}
  }
}
\details{
Heart rates were recorded for a sample of six tree shrews at each of three stages of sleep.
}
\source{
Berger, R. J. and Walker, J. M. (1972) "The Polygraphic Study of Sleep in the Tree Shrew," Brain, Behavior, and Evolution, v. 5, pp. 62
}
\keyword{datasets}
