\name{Stat2Data-package}
\alias{Stat2Data-package}
\alias{Stat2Data}
\docType{package}
\title{
DataSets for Stat2 Textbook
}
\description{
DataSets for Stat2 Textbook (by Cannon, et. al.)
}
\details{
\tabular{ll}{
Package: \tab Stat2Data\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2012-09-11\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
%%
}
\author{
Robin Lock

Maintainer: Robin Lock <rlock@stlawu.edu>
}

\keyword{ package }

