\name{Stat2Data-package}
\alias{Stat2Data-package}
\alias{Stat2Data}
\docType{package}
\title{
DataSets for Stat2 Textbook
}
\description{
DataSets for Stat2 Textbook
}
\details{
\tabular{ll}{
Package: \tab Stat2Data\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-11\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it Robin Lock

Maintainer: Who to complain to <rlock@stlawu.edu>
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
