\name{Stat2Data-package}
\alias{Stat2Data-package}
\alias{Stat2Data}
\docType{package}
\title{
Datasets for Stat2: Modeling with Regression and ANOVA
}
\description{
Datasets for Stat2: Modeling with Regression and ANOVA (second edition) and Stat2: Building Models for a World of Data (first edition)
}
\details{
\tabular{ll}{
Package: \tab Stat2Data\cr
Type: \tab Package\cr
Version: \tab 2.0.0\cr
Date: \tab 2018-12-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
%%
This package included datasets for both the first and second editions of the text.
}
\author{
Ann Cannon, George Cobb, Bradley Hartlaub, Julie Legler, Robin Lock, Thomas Moore, Allan Rossman, Jeffrey Witmer

Maintainer: Robin Lock <rlock@stlawu.edu>
}

\keyword{ package }

