\name{Titanic}
\alias{Titanic}
\docType{data}
\title{Titanic}
\description{
List and outcomes for passengers on the Titanic
}
%\usage{data(Titanic)}
\format{
  A dataset with 1313 observations on the following 6 variables.
  \tabular{rl}{
    high-leverage\code{Name} \tab {Passenger name}\cr
    high-leverage\code{PClass} \tab {Passenger class: \code{*}=missing, \code{1st}, \code{2nd}, or \code{3rd}}\cr
    high-leverage\code{Age} \tab {Age (in years)}\cr
    high-leverage\code{Sex} \tab {\code{female} or \code{male}}\cr
    high-leverage\code{Survived} \tab {\code{1}=survived or \code{0}=died}\cr
    high-leverage\code{SexCode} \tab {\code{1}=female or \code{0}=male}\cr
  }
}
\details{
The Titanic was a British luxury ocean liner that sank famously in the icy North Atlantic 
on its maiden voyage in April of 1912. Of the approximately 2200 passengers on board, 1500
died. The high death rate was blamed largely on the inadequate supply of lifeboats, a result of the
manufacturer's claim that the ship was "unsinkable." A partial data set of the passenger list was
compiled by Philip Hinde in his Encyclopedia Titanica and is given in this dataset.
}
\source{
Philip Hinde's Encyclopedia Titanica, http://www.encyclopedia-titanica.org/.  
Data may also be downloaded from the Australasian Data and Story Library (OzDASL) at
http://www.statsci.org/data/general/titanic.html.
}

\keyword{datasets}
