\name{TomlinsonRush}
\alias{TomlinsonRush}
\docType{data}
\title{
LaDainian Tomlinson Rushing Yards
}
\description{
Rushing yards for each game LaDainian Tomlinson played in the 2006 National Football League (NFL regular) season. 
}
%\usage{data(TomlinsonRush)}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{Game}}{Week number in the 2006 season}
    \item{\code{Opponent}}{Name of opposing team}
    \item{\code{Attempts}}{Number of rushing attempts}
    \item{\code{Yards}}{Total yards gamed rushing for the game}
  }
}
\details{
For each of the sixteen games the San Diego Chargers played in the 2006 NFL regular season 
we have the number of times LaDainian Tomlinson ran the ball and the total yards he gained. 
}
\source{
Data downloaded from http://www.pro-football-reference.com/players/T/TomlLa00/gamelog/2006/
}

\keyword{datasets}
