\name{USstamps}
\alias{USstamps}
\docType{data}
\title{
USstamps
}
\description{
Price of US stamp for first class mail 1885-2012
}
\usage{data(USstamps)}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Years when stamp price changed}
    \item{\code{Price}}{Cost of a US first class stamp (in cents)}
 }
}
\details{
The data record the year and price for each change in price for a 
US first class (1 ounce, domestic letter) stamp since 1885. 
}
\source{
http://about.usps.com/who-we-are/postal-history/domestic-letter-rates-1863-2011.htm
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(USstamps)
}
\keyword{datasets}
