\name{YouthRisk2009}
\alias{YouthRisk2009}
\docType{data}
\title{
YouthRisk2009
}
\description{
Survey of students in grades 9-12 concerning health-risl behiviors
}
\usage{data(YouthRisk2009)}
\format{
  A data frame with 500 observations on the following 6 variables.
  \describe{
    \item{\code{Sleep}}{Average hours sleep oin school night, coded with \code{10 or more hours}, \code{9 hours}, 
\code{8 hours}, \code{7 hours}, \code{6 hours}, \code{5 hours}, or \code{4 or less hours}}
    \item{\code{Sleep7}}{Seven or more hours of sleep? 0=\code{No}, 1=\code{Yes}}
    \item{\code{SmokeLife}}{Ever smoked? \code{No} or \code{Yes}}
    \item{\code{SmokeDaily}}{Regular smoker? \code{No} or \code{Yes}} 
    \item{\code{MarijuaEver}}{Ever smoked marijuana? 0=\code{No} or 1=\code{Yes}}
    \item{\code{Age}}{Age (in years)} 
   }
}
\details{
Data from the Centers for Disease Control's Youth Risk Behavior Surveillance System (YRBSS). 
}
\source{
http://www.cdc.gov/HealthyYouth/yrbs/index.htm
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(YouthRisk2009)
}
\keyword{datasets}
