\name{YouthRisk2009}
\alias{YouthRisk2009}
\docType{data}
\title{Youth Risk Survey}
\description{
Survey of students in grades 9-12 concerning health-risk behiviors
}
\format{
  A dataset with 500 observations on the following 6 variables.
  \describe{
    \item{\code{Sleep}}{Average hours sleep on school night (\code{10 or more hours}, \code{9 hours}, down to
  \code{4 or less hours})}
    \item{\code{Sleep7}}{Seven or more hours of sleep? (\code{0}=no or \code{1}=yes)}
    \item{\code{SmokeLife}}{Ever smoked? (\code{No} or \code{Yes})}
    \item{\code{SmokeDaily}}{Regular smoker? (\code{No} or \code{Yes})}
    \item{\code{MarijuaEver}}{Ever smoked marijuana? (\code{0}=no or \code{1}=yes)}
    \item{\code{Age}}{Age (in years)}
   }
}
\details{
Data from the Centers for Disease Control's Youth Risk Behavior Surveillance System (YRBSS).

This data set is from the first edition, but not used in the second edition.
}
\source{
http://www.cdc.gov/HealthyYouth/yrbs/index.htm
}

\keyword{datasets}
